### 2024-07-01 #205 @MicahGale
MontePy Submission
Submitting Author: @micahgale
All current maintainers: @micahgale, @tjlaboss
Package Name: MontePy
One-Line Description of Package: MontePy is a python library for reading, editing, and writing MCNP input files.
Repository Link:  https://github.com/idaholab/MontePy
Version submitted: 0.2.10
EiC: TBD
Version accepted: TBD

---

[MontePy](https://montepy.org) is a Python library for reading, editing, and writing MCNP input files. MCNP is the [Monte Carlo N-Particle](https://mcnp.lanl.gov/) radiation transport code that supports 37 particle types, and is widely used in Nuclear Engineering, and Medical Physics. MontePy provides an object-oriented interface for MCNP input files. This allows for easy automation of many different tasks for working with MCNP input files. MontePy does not support MCNP output files

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

    - Scientists and engineers who use MCNP and know python are the primary audience. This will be mostly nuclear engineers, and medical physicists. Use cases are:

     - Automating tedious updates of simulation models (e.g., renumbering all materials to merge two models)
     - Automating generating many permutations of the model for optimization, sensitivity analysis, etc.
     - Extracting information from an existing model in a more legible way.

     - There is also [MCNPy](https://github.rpi.edu/NuCoMP/mcnpy), which reports to provide similar features. I have been unable to verify this as I have been unable to install it. MontePy is different by being written purely in python, and not java, and having a publicly accessible github, that anyone can open an issue for.

     - This is a follow on to the presubmission inquiry: #203.
