### 2024-07-02 292dcef
Made test case to replicate #396.
