### 2024-07-02 928562d
Made simpler mechanism to swap parsers without making blank classes.
