### 2024-07-02 44092c6
Started an sdef style parser.
