### 2024-07-02 235edf4
Made a useful debug tool.
