### 2024-07-02 3e11f99
Made parser more reliable and forbid partial success.
