### 2024-07-02 8e04d25
Added doc strings.
