### 2024-07-02 8f17a7b
Removed tally special classes thanks to machinery.
