### 2024-07-02 2c3c3cb
Made progress on making an sdef parser.
