### 2024-07-02 b6b417e
Added sdef bug to changelog.
