### 2024-07-05 idaholab/MontePy#433 @MicahGale
Add reference to Numjuggler in JOSS paper
I just came across the repo Numjuggler: https://github.com/inr-kit/numjuggler?tab=readme-ov-file. We should probably mention them as well for a possible alternative in our JOSS paper submission.

Libraries to Mention:

* [x] PyNE
* [x] Numjuggler
* [x] mckit
   * [x] add reference in readme to being context-free
* [x] MCNP-Input-Reader
* [x] npp_mcnp_plugin??
* [x] mctools (dodu94)
* [x] mc-tools
* [x] CardShardForMCNP
* [x] wig
* [x] PyMCNP
* [x] BEMP_Thesis
* [x] Funz Plug MCNP
* [x] GNP
* [x] map-stp
* [x] MCNP Conversion tools for OpenMC
* [x] MCNP Input Generator
* [x] MCNP syntax highlighting
* [x] MCNP6-HPGE_Detector_Simulation
* [x] Neutronics Material Maker
* [x] NPP_MCNP_Plugin
* [x] Python-bases software tools for MCNP
* [x] rodcal-mcnp
* [x] t4_geom_convert
* [x] vscode_mcnp
* [x] wig
