+++ doc/source/index.rst
@@ -35,6 +35,7 @@ See Also
 * `MontePy PyPI Project <https://pypi.org/project/montepy/>`_
 * `MCNP 6.3 User Manual <https://www.osti.gov/biblio/1889957>`_ DOI: `10.2172/1889957 <https://doi.org/10.2172/1889957>`_
 * `MCNP 6.2 User Manual <https://mcnp.lanl.gov/pdf_files/TechReport_2017_LANL_LA-UR-17-29981_WernerArmstrongEtAl.pdf>`_
+* `MCNP Forum <https://mcnp.discourse.group/>`_
 
 Indices and tables
 ==================
