+++ pyproject.toml
@@ -7,10 +7,10 @@ description = "A library for reading, editing, and writing MCNP input files"
 readme = "README.md"
 requires-python = ">=3.8"
 maintainers = [
-	{name = "Micah Gale", email = "micah.gale@inl.gov"}
+	{name = "Micah Gale", email = "mgale@montepy.org"}
 ]
 authors = [
-	{name = "Micah Gale", email = "micah.gale@inl.gov"},
+	{name = "Micah Gale", email = "mgale@montepy.org"}
 	{name = "Travis Labossiere-Hickman", email = "Travis.LabossiereHickman@inl.gov"},
 	{name = "Brenna Carbno", email="brenna.carbno@inl.gov"}
 ]
