+++ pyproject.toml
@@ -10,7 +10,7 @@ maintainers = [
 	{name = "Micah Gale", email = "mgale@montepy.org"}
 ]
 authors = [
-	{name = "Micah Gale", email = "mgale@montepy.org"}
+	{name = "Micah Gale", email = "mgale@montepy.org"},
 	{name = "Travis Labossiere-Hickman", email = "Travis.LabossiereHickman@inl.gov"},
 	{name = "Brenna Carbno", email="brenna.carbno@inl.gov"}
 ]
