### 2024-07-08 3356e9b
Removed tallies from documentation site.
