+++ doc/source/api/montepy.data_inputs.rst
@@ -24,8 +24,6 @@ montepy.data\_inputs package
    montepy.data_inputs.material_component
    montepy.data_inputs.mode
    montepy.data_inputs.thermal_scattering
-   montepy.data_inputs.tally
-   montepy.data_inputs.tally_multiplier
    montepy.data_inputs.transform
    montepy.data_inputs.universe_input
    montepy.data_inputs.volume
