### 2024-07-08 355a8d7
Fixed sdef parser by making it its mostly own thing.
