+++ montepy/input_parser/data_parser.py
@@ -16,7 +16,7 @@ class DataParser(MCNP_Parser):
     :rtype: SyntaxNode
     """
 
-    debugfile = "parser.out"
+    debugfile = None
 
     @_(
         "introduction",
@@ -180,19 +180,17 @@ class ParamOnlyDataParser(DataParser):
     :rtype: SyntaxNode
     """
 
+    debugfile = None
+
     @_(
         "param_introduction spec_parameters",
     )
     def param_data_input(self, p):
         ret = {}
-        for key, node in p.introduction.nodes.items():
+        for key, node in p.param_introduction.nodes.items():
             ret[key] = node
-        if hasattr(p, "data"):
-            ret["data"] = p.data
-        else:
-            ret["data"] = syntax_node.ListNode("empty data")
-        if hasattr(p, "parameters"):
-            ret["parameters"] = p.parameters
+        if hasattr(p, "spec_parameters"):
+            ret["parameters"] = p.spec_parameters
         return syntax_node.SyntaxNode("data", ret)
 
     @_(
@@ -217,7 +215,6 @@ class ParamOnlyDataParser(DataParser):
         :returns: all parameters
         :rtype: ParametersNode
         """
-        print(list(p))
         if len(p) == 1:
             params = syntax_node.ParametersNode()
             param = p[0]
@@ -227,13 +224,49 @@ class ParamOnlyDataParser(DataParser):
         params.append(param)
         return params
 
-    @_("classifier param_seperator data")
+    @_("spec_classifier param_seperator data")
     def spec_parameter(self, p):
         return syntax_node.SyntaxNode(
-            p.classifier.prefix.value,
+            p.spec_classifier.prefix.value,
             {
-                "classifier": p.classifier,
+                "classifier": p.spec_classifier,
                 "seperator": p.param_seperator,
                 "data": p.data,
             },
         )
+
+    @_(
+        "KEYWORD",
+    )
+    def spec_data_prefix(self, p):
+        return syntax_node.ValueNode(p[0], str)
+
+    @_(
+        "modifier spec_data_prefix",
+        "spec_data_prefix",
+        "spec_classifier NUMBER",
+        "spec_classifier particle_type",
+    )
+    def spec_classifier(self, p):
+        """
+        The classifier of a data input.
+
+        This represents the first word of the data input.
+        E.g.: ``M4``, `IMP:N`, ``F104:p``
+
+        :rtype: ClassifierNode
+        """
+        if hasattr(p, "spec_classifier"):
+            classifier = p.spec_classifier
+        else:
+            classifier = syntax_node.ClassifierNode()
+
+        if hasattr(p, "modifier"):
+            classifier.modifier = syntax_node.ValueNode(p.modifier, str)
+        if hasattr(p, "spec_data_prefix"):
+            classifier.prefix = p.spec_data_prefix
+        if hasattr(p, "NUMBER"):
+            classifier.number = syntax_node.ValueNode(p.NUMBER, int)
+        if hasattr(p, "particle_type"):
+            classifier.particles = p.particle_type
+        return classifier
