### 2024-07-08 873456e
Tried to make ListNode update no padding if values are rammed together.
