+++ tests/test_source.py
@@ -12,4 +12,5 @@ import pytest
 def test_source_parse_and_parrot(line):
     input = Input([line], BlockType.DATA)
     data = parse_data(input)
+    print(repr(data._tree))
     assert data._tree.format() == line
