### 2024-07-09 33098e1
Add never_pad test to list_node format.
