### 2024-07-09 8d35edb
Created system to only flag never_pad during parsing.
