+++ doc/source/api/montepy.data_inputs.transform.rst
@@ -4,5 +4,6 @@ montepy.data\_inputs.transform module
 
 .. automodule:: montepy.data_inputs.transform
    :members:
+   :inherited-members:
    :undoc-members:
    :show-inheritance:
