+++ doc/source/api/montepy.surface_collection.rst
@@ -4,5 +4,6 @@ montepy.surface\_collection module
 
 .. automodule:: montepy.surface_collection
    :members:
+   :inherited-members:
    :undoc-members:
    :show-inheritance:
