+++ doc/source/api/montepy.materials.rst
@@ -4,5 +4,6 @@ montepy.materials module
 
 .. automodule:: montepy.materials
    :members:
+   :inherited-members:
    :undoc-members:
    :show-inheritance:
