+++ doc/source/api/montepy.input_parser.parser_base.rst
@@ -4,5 +4,6 @@ montepy.input\_parser.parser\_base module
 
 .. automodule:: montepy.input_parser.parser_base
    :members:
+   :inherited-members:
    :undoc-members:
    :show-inheritance:
