+++ doc/source/api/montepy.cell.rst
@@ -4,5 +4,6 @@ montepy.cell module
 
 .. autoclass:: montepy.cell.Cell
    :members:
+   :inherited-members:
    :undoc-members:
    :show-inheritance:
