+++ doc/source/api/montepy.input_parser.shortcuts.rst
@@ -4,5 +4,6 @@ montepy.input\_parser.shortcuts module
 
 .. automodule:: montepy.input_parser.shortcuts
    :members:
+   :inherited-members:
    :undoc-members:
    :show-inheritance:
