+++ doc/source/api/montepy.data_inputs.thermal_scattering.rst
@@ -4,5 +4,6 @@ montepy.data\_inputs.thermal\_scattering module
 
 .. automodule:: montepy.data_inputs.thermal_scattering
    :members:
+   :inherited-members:
    :undoc-members:
    :show-inheritance:
