+++ doc/source/api/montepy.data_inputs.data_input.rst
@@ -4,6 +4,7 @@ montepy.data\_inputs.data\_input module
 
 .. automodule:: montepy.data_inputs.data_input
    :members:
+   :inherited-members:
    :undoc-members:
    :show-inheritance:
    :private-members: _class_prefix, _has_number, _has_classifier
