+++ doc/source/api/montepy.data_inputs.lattice_input.rst
@@ -4,5 +4,6 @@ montepy.data\_inputs.lattice\_input module
 
 .. automodule:: montepy.data_inputs.lattice_input
    :members:
+   :inherited-members:
    :undoc-members:
    :show-inheritance:
