+++ doc/source/api/montepy.utilities.rst
@@ -4,5 +4,6 @@ montepy.utilities module
 
 .. automodule:: montepy.utilities
    :members:
+   :inherited-members:
    :undoc-members:
    :show-inheritance:
