### 2024-07-09 50daa80
Pass changelog-test when no code is modified (#439)
* Only fail changelog-test if montepy/ was updated
