### 2024-07-09 b6d3598
Tried new logic for clearing never_pad. Failed.
