### 2024-07-09 #205 @cmarmo
Hi @MicahGale ! Thank you for submitting your package for pyOpenSci review.
Below are the basic checks that your package needs to pass to begin our review.
If some of these are missing, we will ask you to work on them before the review process begins.

---
---

MontePy is in very good shape, congratulations to all the maintainers for your hard work!

Some comments about the checklist above.

- In the documentation about the [installation of a specific version](https://www.montepy.org/starting.html#install-specific-version-for-a-project) : I strongly recommend to remove any manual process creating symbolic links and installing requirements in a folder. Virtual environments should always be preferred and you can install specific versions/tags/branches or commit in them with
```
pip install git+https://github.com/idaholab/MontePy.git@<your-version/tag/branch/commit>
```
- The `README.md` redirects to the documentation for details about the installation: may I suggest to explicitely add the command ```pip install```? This will clarify that the package is available from pypi but not from any conda channel.
- The code of conduct is linked in `contributing.md` (btw thanks for expliciting the contact e-mail) but the file is missing and the link gives a "404 not found".

Minor comments not needed to start the review:
- the documentation generally moves from more readable contents to more technical and detailed contents: I would have put the API at the end of the documentation.
- wearing my data processing engineer hat, I would be grateful to have a safe default option backupping  original files when overwriting ... I was checking basic usage documentation and the behaviour of your writing function really scared me.... :sweat_smile:

Final note: I enjoyed this quote from your documentation 🪄

> Creating a new universe is very straight forward. You just need to initialize it with a new number
