+++ doc/source/starting.rst
@@ -33,7 +33,8 @@ System Wide (for the current user)
    In this case the easiest way to deal with this is to open a teminal inside of `jupyter lab` and to install the package there.
 
 
-#. Install it from `PyPI <https://pypi.org>`_ by running ``pip install --user montepy``.
+#. Install it from `PyPI <https://pypi.org>`_ by running ``pip install montepy``. 
+   You may need to run ``pip install --user montepy`` if you are not allowed to install the package.
 
 Install specific version for a project
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -41,24 +42,8 @@ The best way maybe to setup a project-specific `conda <https://docs.conda.io/pro
 `Mamba <https://mamba.readthedocs.io/en/latest/user_guide/concepts.html>`_, 
 or a `venv <https://docs.python.org/3/library/venv.html>`_ environment.
 The steps for installing inside one of those environments are the same as the previous steps.
+You can specify a specific version from [PyPI](https://pypi.org/project/montepy/) be installed using ``pip install montepy==version``.
 
-Another option is to clone the repository and to use symbolic-links. In this scenario we'll assume that your local
-repository is located at ``~/dev/montepy``, and your project is located at ``~/foo/bar``. 
-
-#. Move to the repository parent folder: ``cd ~/dev``
-#. Clone this repository: ``git clone https://github.com/idaholab/montepy.git`` 
-#. Enter the repository: ``cd montepy``
-#. Checkout the specific version you want. These are tagged with git tags
-
-    #. You can list all tags with ``git tag``
-
-    #. You can then checkout that tag: ``git checkout <tag>``
-
-#. Install the dependent requirements: ``pip install -r requirements/common.txt``
-#. Move to your project folder: ``cd ~/foo/bar``
-#. Create a symbolic link in the project folder to the repository: ``ln -s ~/dev/montepy/montepy montepy``
-
-Now when you run a python script in that folder (*and only in that folder*) ``import montepy`` will use the specific version you want. 
 
 Reading a File
 --------------
