### 2024-07-10 #205 @MicahGale
@cmarmo thank you for the feedback!

I have opened this PR: https://github.com/idaholab/MontePy/pull/440 to address this feedback.

> In the documentation about the [installation of a specific version](https://www.montepy.org/starting.html#install-specific-version-for-a-project) : I strongly recommend to remove any manual process creating symbolic links and installing requirements in a folder. Virtual environments should always be preferred and you can install specific versions/tags/branches or commit in them with

These directions are old and from when this was an internal tool. I just removed them and instead pointed to using `pip install montepy==<version>`. Most versions are on PyPI, and those that aren't, can't be because they don't have an OSS license yet. So if someone really wants that old of a version I think they will figure it out on their own how to manage.

> The README.md redirects to the documentation for details about the installation: may I suggest to explicitely add the command pip install? This will clarify that the package is available from pypi but not from any conda channel.

Added the command.

> The code of conduct is linked in contributing.md (btw thanks for expliciting the contact e-mail) but the file is missing and the link gives a "404 not found".

Corrected this and actually added a boilerplate code of conduct.

> the documentation generally moves from more readable contents to more technical and detailed contents: I would have put the API at the end of the documentation.

Good point. I updated the index to list the API documentation last.

> wearing my data processing engineer hat, I would be grateful to have a safe default option backupping original files when overwriting ... I was checking basic usage documentation and the behaviour of your writing function really scared me.... 😅

This warning was written when MontePy used to discard user formatting and comments, which is no longer the case.
`write_to_file` has no default option that would override the original file.
I think now the only real risk with overwriting the original file is having a script that is buggy and accidentally changing the model.
I changed the warning to discourage this sort of workflow for making script development easier.

> Final note: I enjoyed this quote from your documentation 🪄

>> Creating a new universe is very straight forward. You just need to initialize it with a new number

I forget sometimes about how silly the concept of universes are in these models is sometimes especially when working with them.
