### 2024-07-10 64052ae
Revert "Tried new logic for clearing never_pad. Failed."
This reverts commit b6d3598c44fc261d3b77c411b3ec5fd5be2d4144.
