### 2024-07-10 3ae1bfa
relaxed tests with trailing spaces.
