### 2024-07-10 ce11882
Removed crufty infinite recursion line.
