### 2024-07-10 idaholab/MontePy#442 @MicahGale
Ok I see value in this. I think I'd prefer to just error out rather than creating new files that user wasn't expecting.
