### 2024-07-11 a7d5934
Implemented overwrite guardrails thanks to @tjlaboss.
