### 2024-07-15 11eb1da
Updated dependency to sphinx 7.4.0, and enabled linkcheck testing.
