### 2024-07-15 27e4e77
Switched sphinx process to python 3.12.
