### 2024-07-15 dfaa6f7
Stop supporting python 3.8 for dev work.
