### 2024-07-15 8fe0abc
Updated docstrings to not version changed and errors raised.
