### 2024-07-15 91b8044
Reworded docs to get ready for minor release.
