+++ doc/source/changelog.rst
@@ -1,11 +1,15 @@
 MontePy Changelog
 =================
 
-0.2.10
-----------------------
-**Miscellanious Changes**
+#Next Version#
+---------------------
+
+**Features Added**
 
 * ``overwrite`` argument added to `MCNP_Problem.write_to_file` to ensure files are only overwritten if the user really wants to do so (:pull:`443`).
+
+0.2.10
+----------------------
 
 **Bug fixes**
 
