+++ doc/source/changelog.rst
@@ -6,7 +6,7 @@ MontePy Changelog
 
 **Features Added**
 
-* ``overwrite`` argument added to `MCNP_Problem.write_to_file` to ensure files are only overwritten if the user really wants to do so (:pull:`443`).
+* ``overwrite`` argument added to ``MCNP_Problem.write_to_file`` to ensure files are only overwritten if the user really wants to do so (:pull:`443`).
 
 0.2.10
 ----------------------
