### 2024-07-17 0d27e03
correct next version release number.
