### 2024-07-17 9318acb
Updating changelog to next version: 0.3.0
