### 2024-07-17 df48f1e
Moved to python 3.12 for everything that isn't specifically matrixed.
