+++ .github/workflows/main.yml
@@ -128,10 +128,10 @@ jobs:
     
     steps:
       - uses: actions/checkout@v4
-      - name: set up python 3.8
+      - name: set up python 3.12
         uses: actions/setup-python@v5
         with: 
-          python-version: 3.8
+          python-version: 3.12
       - run: pip install . montepy[format]
       - run: black --check montepy/ tests/
           
