### 2024-07-18 3b81a87
Added testing setuptools_scm to build test.
