### 2024-07-18 c0227a2
Fixed setuptools-scm pypi name.
