### 2024-07-18 c2cbc3c
Add setuptools as dependency to allow setuptools_scm to be happy.
