+++ pyproject.toml
@@ -41,6 +41,7 @@ doc = ["sphinx>=7.4.0", "sphinxcontrib-apidoc", "sphinx_rtd_theme"]
 format = ["black>=23.3.0"]
 build = [
 	"build",
+	"setuptools>=64.0.0",
 	"setuptools-scm>=8",
 ]
 develop = [
@@ -61,7 +62,7 @@ Documentation = "https://idaholab.github.io/MontePy/index.html"
 "change_to_ascii" = "montepy._scripts.change_to_ascii:main"
 
 [build-system]
-requires = ["setuptools  >= 64.0.0", "setuptools_scm>=8"]
+requires = ["setuptools>=64.0.0", "setuptools_scm>=8"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools_scm]
