### 2024-07-18 1bb928f
Made it so setuptools_scm version is displayed.
