+++ .github/workflows/main.yml
@@ -14,6 +14,9 @@ jobs:
 
     steps:
       - uses: actions/checkout@v4
+        with: 
+          fetch-depth: 0
+          fetch-tags: true
       - name: set up python  ${{ matrix.python-version }}
         uses: actions/setup-python@v5
         with: 
@@ -22,7 +25,7 @@ jobs:
       - run: pip install .[build]
       - name: Get Version
         id: get_version
-        run: echo "version=`python -m setuptools_scm`" >> $GITHUB_OUTPUT
+        run: python -m setuptools_scm
       - run: python -m build --sdist --wheel
       - run: pip install .
       - run: pip uninstall -y montepy
