### 2024-07-18 6dd7346
Parameterized the cell init tests.
