### 2024-07-18 36dd12f
Added test to replicate issue #355
