### 2024-07-18 f0c1d77
Parametrized most cell edge case tests.
