+++ .github/workflows/main.yml
@@ -81,13 +81,6 @@ jobs:
         with:
           name: coverage
           path: coverage.xml
-      - name: Test Reporter
-        if: ${{ matrix.python-version == '3.9' && (success() || failure() )}}
-        uses: dorny/test-reporter@v1.7.0
-        with:
-          name: CI-test-report
-          path: test_report.xml
-          reporter: java-junit
       - name: Coveralls GitHub Action
         if: ${{ matrix.python-version == '3.9' && (success() || failure() )}}
         uses: coverallsapp/github-action@v2
