### 2024-07-19 1139eeb
Added parsing rule for implicit intersection.
