+++ montepy/meta.yaml
@@ -1,0 +1,48 @@
+{% set name = "montepy" %}
+{% set version = "0.2.10" %}
+
+package:
+  name: {{ name|lower }}
+  version: {{ version }}
+
+source:
+  url: https://pypi.io/packages/source/{{ name[0] }}/{{ name }}/montepy-{{ version }}.tar.gz
+  sha256: b611fea92fbe8e4088e6550ff81aa65e091478ca7e8cdce4d794ab3ba1ea5a32
+
+build:
+  entry_points:
+    - change_to_ascii = montepy._scripts.change_to_ascii:main
+  noarch: python
+  script: {{ PYTHON }} -m pip install . -vv --no-deps --no-build-isolation
+  number: 0
+
+requirements:
+  host:
+    - python >=3.8
+    - setuptools >=64.0.0
+    - setuptools-scm >=8
+    - pip
+  run:
+    - python >=3.8
+    - numpy >=1.18
+    - sly ==0.5
+
+test:
+  imports:
+    - montepy
+  commands:
+    - pip check
+    - change_to_ascii --help
+  requires:
+    - pip
+
+about:
+  summary: A library for reading, editing, and writing MCNP input files
+  license: MIT AND Unlicense
+  license_file:
+    - LICENSE
+    - NOTICE.txt
+
+extra:
+  recipe-maintainers:
+    - AddYourGitHubIdHere
