+++ doc/source/changelog.rst
@@ -1,6 +1,13 @@
 MontePy Changelog
 =================
 
+#Next Version#
+----------------
+
+**Bug fixes**
+* Fixed parser bug with parsing cells with implicit intersection, e.g., `` (1:-2)(3:-4)``. (:issue:`355`).
+
+
 0.3.0
 -------------------
 
