+++ doc/source/changelog.rst
@@ -5,7 +5,8 @@ MontePy Changelog
 ----------------
 
 **Bug fixes**
-* Fixed parser bug with parsing cells with implicit intersection, e.g., `` (1:-2)(3:-4)``. (:issue:`355`).
+
+* Fixed parser bug with parsing cells with implicit intersection, e.g., ``(1:-2)(3:-4)``. (:issue:`355`).
 
 
 0.3.0
