### 2024-07-19 8bccc0f
Updating changelog to next version: 0.3.1
