### 2024-07-19 #205 @cmarmo
Thank you @MicahGale !
Your package looks ready for review to me.

I noticed that you submitted to JOSS independently (see https://github.com/openjournals/joss-reviews/issues/6977): may I suggest to merge the two submissions, as pyOpenSci has a [partnership with JOSS](https://www.pyopensci.org/software-peer-review/about/intro.html#partnership-with-joss).
This will also lower the pression on our two communities, as both pyOpenSci and JOSS are based on volounteer engagement.
If you are ok with that I can comment on the related JOSS issue and we can follow up here.
