### 2024-07-19 #205 @MicahGale
Yes let's merge them if that makes sense. I just did things in a bit of a different order.
