### 2024-07-22 d7c06ef
Moved only first c style comment for trailing comments.
