### 2024-07-22 ce579b6
Added bug fix to changelog.
