+++ doc/source/changelog.rst
@@ -2,7 +2,7 @@ MontePy Changelog
 =================
 
 #Next Version#
--------------
+--------------
 
 **Bug fixes**
 
