### 2024-07-22 66089ea
Updating changelog to next version: 0.3.2
