### 2024-08-02 9115fd8
Test the addition of a numbered object from another problem
