### 2024-08-02 d28fb0b
In NumberedObjectCollection.append_renumber(), move object linking before the attempt at renumbering.
