### 2024-08-05 a83092d
Assert that material is renumbered, added, and still itself.
