+++ tests/test_numbered_collection.py
@@ -118,8 +118,14 @@ class TestNumberedObjectCollection(unittest.TestCase):
     def test_append_renumber_problems(self):
         prob1 = copy.deepcopy(self.simple_problem)
         prob2 = copy.deepcopy(self.simple_problem)
+        # Delete Material 2, making its number available.
         prob2.materials.remove(prob2.materials[2])
-        prob2.materials.append_renumber(prob1.materials[1])
+        len_mats = len(prob2.materials)
+        mat1 = prob1.materials[1]
+        new_num = prob2.materials.append_renumber(mat1)
+        self.assertEqual(new_num, 2)
+        self.assertEqual(len(prob2.materials), len_mats + 1)
+        self.assertIs(prob2.materials[2], mat1)
 
     def test_request_number(self):
         cells = self.simple_problem.cells
