+++ tests/test_numbered_collection.py
@@ -123,9 +123,9 @@ class TestNumberedObjectCollection(unittest.TestCase):
         len_mats = len(prob2.materials)
         mat1 = prob1.materials[1]
         new_num = prob2.materials.append_renumber(mat1)
-        self.assertEqual(new_num, 2)
-        self.assertEqual(len(prob2.materials), len_mats + 1)
-        self.assertIs(prob2.materials[2], mat1)
+        assert new_num == 2, "Material not renumbered correctly."
+        assert len(prob2.materials) == len_mats + 1, "Material not appended"
+        assert prob2.materials[2] is mat1, "Material 2 is not the new material"
 
     def test_request_number(self):
         cells = self.simple_problem.cells
