### 2024-08-08 c4408c4
Stop relying solely on callers for type detection.
