### 2024-08-08 a67a072
Updated changelog for bug.
