+++ doc/source/changelog.rst
@@ -6,7 +6,8 @@ MontePy Changelog
 
 **Bug fixes**
 
-* Fixed bug with appending and renumbeing numbered objects from other MCNP problems (:issue:`466`).
+* Fixed bug with appending and renumbering numbered objects from other MCNP problems (:issue:`466`).
+* Fixed bug with dynamic typing and the parsers that only appear in edge cases (:issue:`461`).
 
 
 0.3.2
