### 2024-08-08 5c7f2c0
Migrated some material tests to pytest.
