### 2024-08-08 4733b34
Clarified how material_components is built.
