+++ montepy/data_inputs/material.py
@@ -112,6 +112,9 @@ class Material(data_input.DataInputAbstract, Numbered_MCNP_Object):
         """
         The internal dictionary containing all the components of this material.
 
+        The keys are :class:`~montepy.data_inputs.isotope.Isotope` instances, and the values are
+        :class:`~montepy.data_inputs.material_component.MaterialComponent` instances.
+
         :rtype: dict
         """
         return self._material_components
