### 2024-08-08 57b690c
Wrote tests to actually test that the output material is updated.
