### 2024-08-08 0deee48
Update test_isotope_str
