### 2024-08-08 b758397
Update test_material_str
