+++ tests/test_material.py
@@ -98,16 +98,17 @@ class testMaterialClass(TestCase):
         in_str = "M20 1001.80c 0.5 8016.80c 0.4 94239.80c 0.1"
         input_card = Input([in_str], BlockType.DATA)
         material = Material(input_card)
-        answers = """MATERIAL: 20 fractions: atom
- H-1   (80c) 0.5
- O-16  (80c) 0.4
-Pu-239 (80c) 0.1
+        answers = """\
+MATERIAL: 20 fractions: atom
+ZAID=1001, Z=1, A=1, element=hydrogen, library=80c 0.5
+ZAID=8016, Z=8, A=16, element=oxygen, library=80c 0.4
+ZAID=94239, Z=94, A=239, element=plutonium, library=80c 0.1
 """
         output = repr(material)
         print(output)
-        self.assertEqual(output, answers)
+        assert output == answers
         output = str(material)
-        self.assertEqual(output, "MATERIAL: 20, ['hydrogen', 'oxygen', 'plutonium']")
+        assert output == "MATERIAL: 20, ['hydrogen', 'oxygen', 'plutonium']"
 
     def test_material_sort(self):
         in_str = "M20 1001.80c 0.5 8016.80c 0.5"
