### 2024-08-08 812004c
Changed material to only use isotopesNode in parsing.
