### 2024-08-08 5c01053
Run "black" on test files as well.
