+++ tests/test_material.py
@@ -226,8 +226,7 @@ class TestIsotope(TestCase):
         isotope = Isotope("1001.80c")
         assert isotope.mcnp_str() == "1001.80c"
         assert repr(isotope) == "Isotope('1001.80c')"
-        assert str(isotope) == \
-            "ZAID=1001, Z=1, A=1, element=hydrogen, library=80c"
+        assert str(isotope) == "ZAID=1001, Z=1, A=1, element=hydrogen, library=80c"
         isotope = Isotope("94239.80c")
         assert isotope.mcnp_str() == "94239.80c"
         assert repr(isotope) == "Isotope('94239.80c')"
