### 2024-08-08 e98eb45
Simplfied logic to be clearer.
