### 2024-08-09 26ce6c3
Added issue #476 to changelog.
