+++ tests/test_material.py
@@ -122,6 +122,12 @@ Pu-239 (80c) 0.1
         ("M20 1001.80c -0.5 8016.80c -0.5", 20, False, [0.5, 0.5]),
         ("M20 1001.80c -0.5 8016.710nc -0.5", 20, False, [0.5, 0.5]),
         ("M20 1001.80c 0.5 8016.80c 0.5 Gas=1", 20, True, [0.5, 0.5]),
+        (
+            "m1      8016.71c  2.6999999-02 8017.71c  9.9999998-01 plib=84p",
+            1,
+            True,
+            [2.6999999e-2, 9.9999998e-01],
+        ),
     ],
 )
 def test_material_init(line, mat_number, is_atom, fractions):
