### 2024-08-09 1eaedba
Added issue #477 in changelog.
