+++ doc/source/changelog.rst
@@ -8,6 +8,7 @@ MontePy Changelog
 
 * Fixed bug with appending and renumbering numbered objects from other MCNP problems (:issue:`466`).
 * Fixed bug with dynamic typing and the parsers that only appear in edge cases (:issue:`461`).
+* Fixed bug with trying to get trailing comments from non-existant parts of the syntax tree (:pull:`480`).
 
 
 0.3.2
