### 2024-08-09 288f7ca
Enabled making a sitemap.
