+++ doc/source/conf.py
@@ -36,12 +36,15 @@ extensions = [
     "sphinx.ext.napoleon",
     "sphinx.ext.intersphinx",
     "sphinx.ext.extlinks",
+    "sphinx_sitemap",
 ]
 
 # Add any paths that contain templates here, relative to this directory.
 templates_path = ["_templates"]
 html_favicon = "monty.svg"
 html_logo = "monty.svg"
+
+html_baseurl = "https://www.montepy.org/"
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
