+++ doc/source/conf.py
@@ -45,6 +45,7 @@ html_favicon = "monty.svg"
 html_logo = "monty.svg"
 
 html_baseurl = "https://www.montepy.org/"
+html_extra_path = ["robots.txt"]
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This pattern also affects html_static_path and html_extra_path.
