### 2024-08-10 2b83d9f
Update MCNP_str to be more rigurous.
