### 2024-08-10 0e9517f
Fixed keyerror because isotope singleton isn't working.
