### 2024-08-10 0f0299e
Ensured isotope always has a syntax tree.
