### 2024-08-10 5c1c26c
Always update the material syntax tree.
