### 2024-08-11 ef64a35
Made a universal name parser re.
