### 2024-08-11 457c59a
Started making universal isotope parser.
