### 2024-08-12 f5990cb
Correctly implemented string based fancy names for isotopes.
