+++ montepy/data_inputs/isotope.py
@@ -4,6 +4,8 @@ from montepy.data_inputs.element import Element
 from montepy.errors import *
 from montepy.input_parser.syntax_node import ValueNode
 
+import re
+
 
 class Isotope:
     """
@@ -30,26 +32,70 @@ class Isotope:
     )
     """"""
 
-    def __init__(self, ZAID="", element=None, Z=None, A=None, node=None):
+    def __init__(
+        self,
+        ZAID="",
+        element=None,
+        Z=None,
+        A=None,
+        meta_state=None,
+        library="",
+        node=None,
+    ):
         if node is not None and isinstance(node, ValueNode):
             if node.type == float:
                 node = ValueNode(node.token, str, node.padding)
             self._tree = node
             ZAID = node.value
-        parts = ZAID.split(".")
-        try:
-            assert len(parts) <= 2
-            int(parts[0])
-        except (AssertionError, ValueError) as e:
-            raise ValueError(f"ZAID: {ZAID} could not be parsed as a valid isotope")
-        self._ZAID = parts[0]
-        self.__parse_zaid()
-        if len(parts) == 2:
-            self._library = parts[1]
+        if ZAID:
+            parts = ZAID.split(".")
+            try:
+                assert len(parts) <= 2
+                int(parts[0])
+            except (AssertionError, ValueError) as e:
+                raise ValueError(f"ZAID: {ZAID} could not be parsed as a valid isotope")
+            self._ZAID = parts[0]
+            new_vals = self._parse_zaid(int(self._ZAID))
+            for key, value in new_vals.items():
+                setattr(self, key, value)
+            if len(parts) == 2:
+                self._library = parts[1]
+            else:
+                self._library = ""
+            return
+        elif element is not None:
+            if not isinstance(element, Element):
+                raise TypeError(
+                    f"Only type Element is allowed for element argument. {element} given."
+                )
+            self._element = element
+            self._Z = self._element.Z
+        elif Z is not None:
+            if not isinstance(Z, int):
+                raise TypeError(f"Z number must be an int. {Z} given.")
+            self._Z = Z
+            self._element = Element(Z)
+        if A is not None:
+            if not isinstance(A, int):
+                raise TypeError(f"A number must be an int. {A} given.")
+            self._A = A
         else:
-            self._library = ""
+            self._A = 0
+        if not isinstance(meta_state, (int, type(None))):
+            raise TypeError(f"Meta state must be an int. {meta_state} given.")
+        if meta_state:
+            self._is_metastable = True
+            self._meta_state = meta_state
+        else:
+            self._is_metastable = False
+            self._meta_state = None
+        if not isinstance(library, str):
+            raise TypeError(f"Library can only be str. {library} given.")
+        self._library = library
+        self._ZAID = str(self.get_full_zaid())
 
-    def __parse_zaid(self):
+    @classmethod
+    def _parse_zaid(cls, ZAID):
         """
         Parses the ZAID fully including metastable isomers.
 
@@ -58,7 +104,7 @@ class Isotope:
         """
 
         def is_probably_an_isotope(Z, A):
-            for lim_Z, lim_A in self._BOUNDING_CURVE:
+            for lim_Z, lim_A in cls._BOUNDING_CURVE:
                 if Z <= lim_Z:
                     if A <= lim_A:
                         return True
@@ -69,24 +115,24 @@ class Isotope:
             # if you are above Lv it's probably legit.
             return True
 
-        ZAID = int(self._ZAID)
-        self._Z = int(ZAID / 1000)
-        self._element = Element(self.Z)
+        ret = {}
+        ret["_Z"] = int(ZAID / 1000)
+        ret["_element"] = Element(ret["_Z"])
         A = int(ZAID % 1000)
-        if not is_probably_an_isotope(self.Z, A):
-            self._is_metastable = True
[…]
