### 2024-08-12 e0b7b95
Wrote test cases to verify fancy names.
