+++ tests/test_material.py
@@ -251,6 +251,9 @@ class TestIsotope(TestCase):
         ((92, 235, 1, "80c"), 92, 235, 1, "80c"),
         ((Element(92), 235, 1, "80c"), 92, 235, 1, "80c"),
         ((Element(92), 235), 92, 235, None, ""),
+        (("U", 235), 92, 235, None, ""),
+        ((92, 235), 92, 235, None, ""),
+        (("uRanium", 235), 92, 235, None, ""),
         ((Element(92),), 92, 0, None, ""),
     ],
 )
