### 2024-08-12 4b72c02
Handled the case of using a tuple as an isotope identifier.
