### 2024-08-12 5b9dac6
Prototyped the how get/set should work for materials.
