### 2024-08-12 0cdc143
Change Isotope.__repr__()
