### 2024-08-12 bb7e8c1
Started writing tests for the universe moving bug.
