### 2024-08-13 e405498
Made it possible to get material component by just an Isotope.
