+++ tests/test_material.py
@@ -248,6 +248,7 @@ class TestIsotope(TestCase):
         ("hydrogen1m3", 1, 1, 3, ""),
         ("hydrogen1m3.80c", 1, 1, 3, "80c"),
         ("92635m2.710nc", 92, 235, 3, "710nc"),
+        (Isotope("1001.80c"), 1, 1, None, "80c"),
         ((92, 235, 1, "80c"), 92, 235, 1, "80c"),
         ((Element(92), 235, 1, "80c"), 92, 235, 1, "80c"),
         ((Element(92), 235), 92, 235, None, ""),
