### 2024-08-13 935c96e
Made wrapper class for isotope library.
