### 2024-08-13 26ffeb0
Made test to replicate #483.
