### 2024-08-13 080f8ed
Isotope.nuclide_str() when no xs lib
