### 2024-08-13 42cb025
Restore old Isotope __str__
