### 2024-08-13 20a664d
Consider no library possibility in Isotope.mcnp_str()
