### 2024-08-13 9a1f9f6
Test Isotope.nuclide_str()
