### 2024-08-13 a280fb3
Updating changelog to next version: 0.3.3
