### 2024-08-13 51d0ebc
Updated tests to verify #486.
