+++ tests/test_material.py
@@ -15,7 +15,6 @@ from montepy.input_parser.mcnp_input import Input
 
 
 class testMaterialClass(TestCase):
-
     def test_material_parameter_parsing(self):
         for line in ["M20 1001.80c 1.0 gas=0", "M20 1001.80c 1.0 gas = 0 nlib = 00c"]:
             input = Input([line], BlockType.CELL)
@@ -153,7 +152,6 @@ def test_material_update_format():
     ],
 )
 def test_material_init(line, mat_number, is_atom, fractions):
-
     input = Input([line], BlockType.DATA)
     material = Material(input)
     assert material.number == mat_number
@@ -238,7 +236,17 @@ class TestIsotope(TestCase):
         assert isotope.nuclide_str() == "Pu-239.80c"
         assert isotope.mcnp_str() == "94239.80c"
         assert repr(isotope) == "Isotope('Pu-239.80c')"
+        isotope = Isotope("92635.80c")
+        assert isotope.nuclide_str() == "U-235m1.80c"
+        assert isotope.mcnp_str() == "92635.80c"
+        assert str(isotope) == " U-235m1 (80c)"
+        assert repr(isotope) == "Isotope('U-235m1.80c')"
+        # stupid legacy stupidity #486
         isotope = Isotope("95642")
+        assert isotope.nuclide_str() == "Am-242"
+        assert isotope.mcnp_str() == "95642"
+        assert repr(isotope) == "Isotope('Am-242')"
+        isotope = Isotope("95242")
         assert isotope.nuclide_str() == "Am-242"
         assert isotope.mcnp_str() == "95642"
         assert repr(isotope) == "Isotope('Am-242')"
