### 2024-08-13 91a5dbd
Updated isotope str to include metastable.
