### 2024-08-13 #205 @cmarmo
Hello @MicahGale , I'm glad to announce that @kellyrowland have accepted to be editor for the MontePy review.
Thank you so much Kelly!

I'm letting her introduce herself here and I wish to all of you a happy review process! :rocket: :
