### 2024-08-13 cc18df0
Handled stupid legacy edge case.
