### 2024-08-13 c4a3f31
Added issue #486 to changelog.
