### 2024-08-14 a9c6f2f
Played around with changing setattr
