### 2024-08-14 d868dcf
Added attribute error exception.
