+++ tests/test_edge_cases.py
@@ -194,7 +194,8 @@ def test_universe_after_comment():
     problem = montepy.read_input(
         os.path.join("tests", "inputs", "test_universe_data.imcnp")
     )
-    problem.print_in_data_block["u"] = True
+    problem.print_in_data_block["u"] = False
+    problem.print_in_data_block["vol"] = False
     universes = [cell.universe.number for cell in problem.cells]
     try:
         out_file = "universe_after_comment"
@@ -203,8 +204,9 @@ def test_universe_after_comment():
         found = False
         with open(out_file, "r") as fh:
             for line in fh:
+                print(line.rstrip())
                 if "U=" in line:
-                    found == True
+                    found = True
         assert found
         problem = montepy.read_input(out_file)
         new_universes = [cell.universe.number for cell in problem.cells]
