### 2024-08-15 e12485a
Tried to make it so that cell modifiers check for graveyard comments.
