### 2024-08-15 53637c8
Made method to auto insert new materials to data_inputs.
