+++ montepy/materials.py
@@ -1,11 +1,11 @@
 # Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
 import montepy
-from montepy.numbered_object_collection import NumberedObjectCollection
+from montepy.numbered_object_collection import NumberedDataObjectCollection
 
 Material = montepy.data_inputs.material.Material
 
 
-class Materials(NumberedObjectCollection):
+class Materials(NumberedDataObjectCollection):
     """
     A container of multiple :class:`~montepy.data_inputs.material.Material` instances.
 
