### 2024-08-15 b1c889b
Made append method more robust.
