### 2024-08-15 524acf0
Updated changelog with materials append feature.
