+++ doc/source/changelog.rst
@@ -4,6 +4,10 @@ MontePy Changelog
 #Next Version#
 --------------
 
+**Features Added**
+
+* When adding a material to problem.materials it will also be added to problem.data_inputs, ensuring it is printed to the file (:pull:`488`).
+
 **Bug Fixes**
 
 * Fixed bug that didn't show metastable states for pretty printing and isotope. Also handled the case that Am-241 metstable states break convention (:issue:`486`).
