### 2024-08-16 #205 @kellyrowland
Hi -

This is my first engagement with pyOpenSci, so thanks in advance for your patience. 😅 I've been an editor for JOSS for a few years, and that's how we've arrived here.

@MicahGale before I get started on finding reviewers, I see there are a set of JOSS-related boxes to tick off - can you take a look at those and check them off/open PRs/etc. and let me know about the status of those items?

Thanks for tagging some possible reviewers over in https://github.com/openjournals/joss-reviews/issues/6977 - I'll ping folks in this issue and make a post with the editor template once we've got two reviewers on board.

-Kelly
