### 2024-08-16 #205 @MicahGale
Thank you for being willing to do this new role for this package. :)

Ok I updated the JOSS section accordingly.

The one concern I had was about getting a DOI for archiving the software.

Under the JOSS guidelines it seems like that's a final step?

> Upon successful completion of the review, authors will make a tagged release of the software, and deposit a copy of the repository with a data-archiving service such as [Zenodo](https://zenodo.org/) or [figshare](https://figshare.com/), get a DOI for the archive, and update the review issue thread with the version number and DOI.

Are you alright with following the JOSS order for this?
