### 2024-08-16 #205 @kellyrowland
Good point, thanks. I think archiving the release and getting a DOI is a logical last step since it's often the case that changes are made to the software during the review process.
