### 2024-08-16 7550970
Separate writing function from write_to_file()
