### 2024-08-16 db62c4c
Non MCNP_InputFile logic in write_to_file()
