### 2024-08-16 6b118c6
Implement MCNP_Problem.write_problem()
