### 2024-08-16 78abee3
getattr(), not .get()
