+++ doc/source/conf.py
@@ -45,7 +45,6 @@ html_favicon = "monty.svg"
 html_logo = "monty.svg"
 
 html_baseurl = "https://www.montepy.org/"
-sitemap_url_scheme = "{link}"
 html_extra_path = ["robots.txt"]
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
