### 2024-08-17 f427b86
Add sitemap check to CI.
