+++ .github/workflows/main.yml
@@ -116,7 +116,7 @@ jobs:
              cd doc
              make linkcheck
       - name: Test for broken sitemap URLs
-        run: python .github/scripts/check_sitemap.xml
+        run: python .github/scripts/check_sitemap.py 
 
   format-test:
     runs-on: ubuntu-latest
