### 2024-08-17 a40fc71
Moved sitemap test to build step.
