+++ .github/workflows/main.yml
@@ -101,11 +101,8 @@ jobs:
         with: 
           python-version: 3.12
       - run: pip install . montepy[doc,build]
-      - run: |
-             sphinx-build doc/source/ doc/build/ -W --keep-going -E
-             ls doc/build/html
-             python .github/scripts/check_sitemap.py 
-        name: Build site strictly and check sitemap
+      - run: sphinx-build doc/source/ doc/build/html -W --keep-going -E
+        name: Build site strictly
       - uses: actions/upload-artifact@v4
         with:
           name: website
@@ -118,6 +115,8 @@ jobs:
         run: |
              cd doc
              make linkcheck
+      - name: test sitemap
+        run: python .github/scripts/check_sitemap.py 
 
   format-test:
     runs-on: ubuntu-latest
