### 2024-08-17 d0397bf
Added auto deleter.
