### 2024-08-17 1c3d5b0
Added all deleter methods to data numbered.
