### 2024-08-17 29900d0
Remove comments about .tell()
