### 2024-08-17 bc900b0
Directly create MCNP_InputFile from stream
