+++ montepy/input_parser/input_file.py
@@ -31,6 +31,19 @@ class MCNP_InputFile:
         self._overwrite = overwrite
         self._mode = None
         self._fh = None
+    
+    @classmethod
+    def from_open_stream(cls, fh):
+        """
+        Create an MCNP Input File from an open, writable stream
+        
+        :param fh: An open and writable object, such as a file handle.
+        :type fh: io.TextIOBase
+        """
+        name = getattr(fh, "name", fh.__class__.__name__)
+        inpfile = cls(path=name)
+        inpfile._fh = fh
+        return inpfile
 
     @make_prop_pointer("_path")
     def path(self):
