### 2024-08-17 1044fc2
Make write_to_stream() private and deprecate write_to_file()
