### 2024-08-17 a56f294
Improve parameter type hinting
