### 2024-08-18 faa9184
Move test_integration.py to pytest
I also fixed a few minor mistakes throughout
