+++ tests/test_integration.py
@@ -1,1048 +1,1034 @@
 # Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
 import copy
-import unittest
-from unittest import TestCase, expectedFailure
+import pytest
 import os
 
 import montepy
-from montepy.data_inputs import material, thermal_scattering, volume
+from montepy.data_inputs import material, volume
 from montepy.input_parser.mcnp_input import (
     Input,
     Jump,
     Message,
     Title,
-    ReadInput,
 )
 from montepy.errors import *
 from montepy.particle import Particle
 import numpy as np
 
 
-class testFullFileIntegration(TestCase):
-    @classmethod
-    def setUpClass(cls):
-        file_name = "tests/inputs/test.imcnp"
-        cls.simple_problem = montepy.read_input(file_name)
-        cls.importance_problem = montepy.read_input(
-            os.path.join("tests", "inputs", "test_importance.imcnp")
-        )
-        cls.universe_problem = montepy.read_input(
-            os.path.join("tests", "inputs", "test_universe.imcnp")
-        )
+@pytest.fixture
+def simple_problem():
+    return montepy.read_input(
+        os.path.join("tests", "inputs", "test.imcnp")
+    )
 
-    def test_original_input(self):
-        cell_order = [Message, Title] + [Input] * 17
-        for i, input_ob in enumerate(self.simple_problem.original_inputs):
-            self.assertIsInstance(input_ob, cell_order[i])
+@pytest.fixture
+def importance_problem():
+    return montepy.read_input(
+        os.path.join("tests", "inputs", "test_importance.imcnp")
+    )
 
-    def test_original_input_dos(self):
-        problem = montepy.read_input(os.path.join("tests", "inputs", "test_dos.imcnp"))
-        cell_order = [Message, Title] + [Input] * 16
-        for i, input_ob in enumerate(problem.original_inputs):
-            self.assertIsInstance(input_ob, cell_order[i])
+@pytest.fixture
+def universe_problem():
+    return montepy.read_input(
+        os.path.join("tests", "inputs", "test_universe.imcnp")
+    )
 
-    def test_original_input_tabs(self):
-        problem = montepy.read_input(os.path.join("tests", "inputs", "test_tab.imcnp"))
-        cell_order = [Message, Title] + [Input] * 17
-        for i, input_ob in enumerate(problem.original_inputs):
-            self.assertIsInstance(input_ob, cell_order[i])
+@pytest.fixture
+def data_universe_problem():
+    return montepy.read_input(
+        os.path.join("tests", "inputs", "test_universe_data.imcnp")
+    )
 
-    # TODO formalize this or see if this is covered by other tests.
-    def test_lazy_comments_check(self):
-        problem = self.simple_problem
-        material = problem.materials[2]
-        for comment in material._tree.comments:
-            print(repr(comment))
-        print(material._tree.get_trailing_comment())
 
-    def test_material_parsing(self):
-        mat_numbers = [1, 2, 3]
-        for i, mat in enumerate(self.simple_problem.materials):
-            self.assertEqual(mat.number, mat_numbers[i])
+def test_original_input(simple_problem):
+    cell_order = [Message, Title] + [Input] * 17
+    for i, input_ob in enumerate(simple_problem.original_inputs):
+        assert isinstance(input_ob, cell_order[i])
 
-    def test_surface_parsing(self):
-        surf_numbers = [1000, 1005, 1010]
-        for i, surf in enumerate(self.simple_problem.surfaces):
-            self.assertEqual(surf.number, surf_numbers[i])
+def test_original_input_dos():
+    dos_problem = montepy.read_input(
+        os.path.join("tests", "inputs", "test_dos.imcnp")
+    )
+    cell_order = [Message, Title] + [Input] * 16
[…]
