### 2024-08-18 883f8e4
Test new writing methods
