+++ tests/test_integration.py
@@ -1,5 +1,7 @@
 # Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
 import copy
+import io
+
 import pytest
 import os
 
@@ -127,7 +129,8 @@ def test_write_to_file(simple_problem):
     out = "foo.imcnp"
     try:
         problem = copy.deepcopy(simple_problem)
-        problem.write_to_file(out)
+        with pytest.deprecated_call():
+            problem.write_to_file(out)
         with open(out, "r") as fh:
             for line in fh:
                 print(line.rstrip())
@@ -466,111 +469,91 @@ def test_importance_format_mutated(importance_problem):
     assert "imp:n 0.5 1 1 0 3" in output
 
 def test_importance_write_unmutated(importance_problem):
-    out_file = "test_import_unmute"
-    try:
-        importance_problem.write_to_file(out_file)
-        found_np = False
-        found_e = False
-        with open(out_file, "r") as fh:
-            for line in fh:
-                print(line.rstrip())
-                if "imp:n,p 1" in line:
-                    found_np = True
-                elif "imp:e" in line:
-                    found_e = True
-        assert found_np
-        assert found_e
-    finally:
-        try:
-            os.remove(out_file)
-        except FileNotFoundError:
-            pass
+    fh = io.StringIO()
+    importance_problem.write_problem(fh)
+    found_np = False
+    found_e = False
+    fh.seek(0)
+    for line in fh.readlines():
+        print(line.rstrip())
+        if "imp:n,p 1" in line:
+            found_np = True
+        elif "imp:e" in line:
+            found_e = True
+    assert found_np
+    assert found_e
+    fh.close()
 
 def test_importance_write_mutated(importance_problem):
-    out_file = "test_import_mute"
+    fh = io.StringIO()
     problem = copy.deepcopy(importance_problem)
     problem.cells[1].importance.neutron = 0.5
-    try:
-        with pytest.warns(LineExpansionWarning):
-            problem.write_to_file(out_file)
-        found_n = False
-        found_e = False
-        with open(out_file, "r") as fh:
-            for line in fh:
-                print(line.rstrip())
-                if "imp:n 0.5" in line:
-                    found_n = True
-                elif "imp:e" in line:
-                    found_e = True
-        assert found_n
-        assert found_e
-    finally:
-        try:
-            os.remove(out_file)
-        except FileNotFoundError:
-            pass
+    with pytest.warns(LineExpansionWarning):
+        problem.write_problem(fh)
+    found_n = False
+    found_e = False
+    fh.seek(0)
+    for line in fh.readlines():
+        print(line.rstrip())
+        if "imp:n 0.5" in line:
+            found_n = True
+        elif "imp:e" in line:
+            found_e = True
+    assert found_n
+    assert found_e
+    fh.close()
 
 def test_importance_write_cell(importance_problem):
     for state in ["no change", "new unmutated cell", "new mutated cell"]:
-        out_file = "test_import_cell"
+        fh = io.StringIO()
         problem = copy.deepcopy(importance_problem)
         if "new" in state:
             cell = copy.deepcopy(problem.cells[5])
             cell.number = 999
             problem.cells.append(cell)
         problem.print_in_data_block["imp"] = False
-        try:
-            if "new" in state:
-                with pytest.warns(LineExpansionWarning):
-                    problem.write_to_file(out_file)
-            else:
-                problem.write_to_file(out_file)
-            found_np = False
-            found_e = False
-            found_data_np = False
-            with open(out_file, "r") as fh:
-                for line in fh:
-                    print(line.rstrip())
-                    if "imp:n,p=1" in line:
-                        found_np = True
-                    elif "imp:e=1" in line:
[…]
