+++ tests/test_integration.py
@@ -20,21 +20,18 @@ import numpy as np
 
 @pytest.fixture
 def simple_problem():
-    return montepy.read_input(
-        os.path.join("tests", "inputs", "test.imcnp")
-    )
+    return montepy.read_input(os.path.join("tests", "inputs", "test.imcnp"))
+
 
 @pytest.fixture
 def importance_problem():
-    return montepy.read_input(
-        os.path.join("tests", "inputs", "test_importance.imcnp")
-    )
+    return montepy.read_input(os.path.join("tests", "inputs", "test_importance.imcnp"))
+
 
 @pytest.fixture
 def universe_problem():
-    return montepy.read_input(
-        os.path.join("tests", "inputs", "test_universe.imcnp")
-    )
+    return montepy.read_input(os.path.join("tests", "inputs", "test_universe.imcnp"))
+
 
 @pytest.fixture
 def data_universe_problem():
@@ -48,19 +45,20 @@ def test_original_input(simple_problem):
     for i, input_ob in enumerate(simple_problem.original_inputs):
         assert isinstance(input_ob, cell_order[i])
 
+
 def test_original_input_dos():
-    dos_problem = montepy.read_input(
-        os.path.join("tests", "inputs", "test_dos.imcnp")
-    )
+    dos_problem = montepy.read_input(os.path.join("tests", "inputs", "test_dos.imcnp"))
     cell_order = [Message, Title] + [Input] * 16
     for i, input_ob in enumerate(dos_problem.original_inputs):
         assert isinstance(input_ob, cell_order[i])
+
 
 def test_original_input_tabs():
     problem = montepy.read_input(os.path.join("tests", "inputs", "test_tab.imcnp"))
     cell_order = [Message, Title] + [Input] * 17
     for i, input_ob in enumerate(problem.original_inputs):
         assert isinstance(input_ob, cell_order[i])
+
 
 # TODO formalize this or see if this is covered by other tests.
 def test_lazy_comments_check(simple_problem):
@@ -69,15 +67,18 @@ def test_lazy_comments_check(simple_problem):
         print(repr(comment))
     print(material2._tree.get_trailing_comment())
 
+
 def test_material_parsing(simple_problem):
     mat_numbers = [1, 2, 3]
     for i, mat in enumerate(simple_problem.materials):
         assert mat.number == mat_numbers[i]
 
+
 def test_surface_parsing(simple_problem):
     surf_numbers = [1000, 1005, 1010]
     for i, surf in enumerate(simple_problem.surfaces):
         assert surf.number == surf_numbers[i]
+
 
 def test_data_card_parsing(simple_problem):
     M = material.Material
@@ -90,6 +91,7 @@ def test_data_card_parsing(simple_problem):
             assert isinstance(card, cards[i])
         if i == 2:
             assert card.thermal_scattering is not None
+
 
 def test_cells_parsing_linking(simple_problem):
     cell_numbers = [1, 2, 3, 99, 5]
@@ -106,14 +108,17 @@ def test_cells_parsing_linking(simple_problem):
         assert surfaces.union(surf_answer[i]) == surfaces
         assert set(cell.complements).union(complements[i]) == complements[i]
 
+
 def test_message(simple_problem):
     lines = ["this is a message", "it should show up at the beginning", "foo"]
     for i, line in enumerate(simple_problem.message.lines):
         assert line == lines[i]
 
+
 def test_title(simple_problem):
     answer = "MCNP Test Model for MOAA"
     assert answer == simple_problem.title.title
+
 
 def test_read_card_recursion():
     problem = montepy.read_input("tests/inputs/testReadRec1.imcnp")
[…]
