+++ montepy/mcnp_problem.py
@@ -415,7 +415,7 @@ class MCNP_Problem:
     def write_to_file(self, file_path, overwrite=False):
         """
         Writes the problem to a file.
-        
+
         .. versionchanged:: 0.4.0
             Deprecated in favor of self.write_problem()
 
