+++ doc/source/changelog.rst
@@ -1,12 +1,16 @@
 MontePy Changelog
 =================
 
-#Next Version#
---------------
-
+0.4.0
+-----
 **Bug Fixes**
 
 * Fixed bug that didn't show metastable states for pretty printing and isotope. Also handled the case that Am-241 metstable states break convention (:issue:`486`).
+
+**Features Added**
+
+* Write problems to either file paths or streams (file handles) with MCNP_Problem.write_problem() (:issue:`492`).
+
 
 0.3.3
 --------------
