### 2024-08-18 f74b19d
Update edge case test to use write_problem()
