### 2024-08-19 idaholab/MontePy#433 @MicahGale
Just do a check of: https://github.com/topics/mcnp again.
