### 2024-08-19 4adaf1c
use pytest.raises in test_orphaning_mt
