### 2024-08-19 6ef5274
Convert test_mcnp_problem.py to pytest
