### 2024-08-19 363c469
Test catching of writing to unwritable destination
