### 2024-08-19 5460cb2
Set fixture scope to "module" and format w black
