+++ tests/test_mcnp_problem.py
@@ -5,13 +5,15 @@ import montepy
 from montepy.mcnp_problem import MCNP_Problem
 
 
-@pytest.fixture
+@pytest.fixture(scope="module")
 def problem_path():
     return "tests/inputs/test.imcnp"
 
-@pytest.fixture
+
+@pytest.fixture(scope="module")
 def problem(problem_path):
     return MCNP_Problem(problem_path)
+
 
 def test_problem_init(problem, problem_path):
     assert isinstance(
@@ -21,8 +23,10 @@ def test_problem_init(problem, problem_path):
     assert problem.input_file.name == problem_path
     assert problem.mcnp_version == (6, 2, 0)
 
+
 def test_problem_str(problem, problem_path):
     assert f"MCNP problem for: {problem_path}" in str(problem)
+
 
 def test_problem_write_type(problem):
     unwritable = object()
