+++ tests/test_mcnp_problem.py
@@ -27,6 +27,9 @@ def test_problem_init(problem, problem_path):
 def test_problem_str(problem, problem_path):
     assert f"MCNP problem for: {problem_path}" in str(problem)
 
+def test_problem_repr(problem, problem_path):
+    assert repr(problem).startswith(f"MCNP problem for: {problem_path}")
+
 
 def test_problem_write_type(problem):
     unwritable = object()
