### 2024-08-19 74d415a
Bring changelog up to standards
