### 2024-08-19 idaholab/MontePy#433 @MicahGale
Also review: https://github.com/MCNP-Devel
