### 2024-08-19 3c6f648
Switched to stringIO.
