### 2024-08-20 bd9215b
Added #483 to the changelog.
