### 2024-08-20 bcc8d32
Handled case of None in a syntax tree.
