### 2024-08-20 5148212
Ensured that format_for_mcnp doesn't strip out necessary blank lines for cell_modifiers.
