+++ montepy/mcnp_object.py
@@ -214,21 +214,20 @@ class MCNP_Object(ABC):
         )
         ret = []
         for line in strings:
-            if line.strip():
-                buffer = wrapper.wrap(line)
-                if len(buffer) > 1:
-                    warning = LineExpansionWarning(
-                        f"The line exceeded the maximum length allowed by MCNP, and was split. The line was:\n{line}"
-                    )
-                    warning.cause = "line"
-                    warning.og_value = line
-                    warning.new_value = buffer
-                    warnings.warn(
-                        warning,
-                        LineExpansionWarning,
-                        stacklevel=2,
-                    )
-                ret += buffer
+            buffer = wrapper.wrap(line)
+            if len(buffer) > 1:
+                warning = LineExpansionWarning(
+                    f"The line exceeded the maximum length allowed by MCNP, and was split. The line was:\n{line}"
+                )
+                warning.cause = "line"
+                warning.og_value = line
+                warning.new_value = buffer
+                warnings.warn(
+                    warning,
+                    LineExpansionWarning,
+                    stacklevel=2,
+                )
+            ret += buffer
         return ret
 
     def validate(self):
