### 2024-08-20 56c4dd0
Ensured that cell modifier blank lines don't get left in.
