### 2024-08-20 991a79c
Added doc strings for new functions.
