+++ tests/test_syntax_parsing.py
@@ -567,6 +567,7 @@ class TestPaddingNode(TestCase):
         self.assertEqual(len(list(comment.comments)), 1)
         self.assertEqual(len(comment.contents), 0)
 
+
 def test_graveyard_comment():
     padding = syntax_node.PaddingNode(" ")
     padding.append("$ test", True)
@@ -1206,8 +1207,7 @@ test title
         card = montepy.input_parser.syntax_node.CommentNode(in_strs[0])
         output = card.format()
         answer = "c foo"
-        str_answer = """COMMENT:
-c foo"""
+        str_answer = "COMMENT: c foo"
         self.assertEqual(repr(card), str_answer)
         self.assertEqual("c foo", str(card))
         self.assertEqual(len(answer), len(output))
