### 2024-08-20 f72e9db
Fixed issue with geometry tree being excluded from flat packing.
