### 2024-08-20 30d92ea
Replicated another edge case for travis.
