### 2024-08-20 0ade822
Fixed bug with cell_modifiers being printed in comments.  (#484)
* Started writing tests for the universe moving bug.

* Made test to replicate #483.

* Confused myself about what true means.

* Tried to implement avoiding graveyard comments.

* added unit test for graveyard comment checker.

* Tried to make it so that cell modifiers check for graveyard comments.

* Added #483 to the changelog.

* Handled case of None in a syntax tree.

* Ensured that format_for_mcnp doesn't strip out necessary blank lines for cell_modifiers.

* Ensured that cell modifier blank lines don't get left in.

* Adde
[…]
