+++ tests/test_edge_cases.py
@@ -183,3 +183,33 @@ def test_complex_cell_parsing(lines):
         montepy.input_parser.block_type.BlockType.CELL,
     )
     montepy.Cell(input)
+
+
+@pytest.mark.filterwarnings("ignore")
+def test_universe_after_comment():
+    problem = montepy.read_input(
+        os.path.join("tests", "inputs", "test_universe_data.imcnp")
+    )
+    problem.print_in_data_block["u"] = False
+    problem.print_in_data_block["vol"] = False
+    universes = [cell.universe.number for cell in problem.cells]
+    try:
+        out_file = "universe_after_comment"
+        problem.write_to_file(out_file)
+        # TODO check that the Universe does show up
+        found = False
+        with open(out_file, "r") as fh:
+            for line in fh:
+                print(line.rstrip())
+                assert "c IMP:n=0.0" not in line
+                if "U=" in line:
+                    found = True
+        assert found
+        problem = montepy.read_input(out_file)
+        new_universes = [cell.universe.number for cell in problem.cells]
+        assert universes == new_universes
+    finally:
+        try:
+            os.remove(out_file)
+        except FileNotFoundError:
+            pass
