### 2024-08-21 9a97f23
Stoped hiding the shift operator.
