### 2024-08-22 15752ae
Implemented updating syntax trees for needed parentheses.
