### 2024-08-22 28033c1
Added Operator.GROUP for clarity.
