+++ montepy/geometry_operators.py
@@ -23,6 +23,9 @@ class Operator(Enum):
     """
     Internal operator essentially equivalent to No-op.
 
-    This is used to properly handle parentheses while parsing.
+    This is used to properly handle some leaf nodes.
     """
-    # TODO decide on group
+    GROUP = "()"
+    """
+    Grouping operator that represents parentheses.
+    """
