### 2024-08-22 #205 @kellyrowland
hi @paulromano @munkm 👋 would you be interested in and available to review this pyOpenSci submission?

the reviewer template that you would use can be seen at https://www.pyopensci.org/software-peer-review/appendices/templates.html#peer-review-template .

if you're not available for the review, could you suggest other potential reviewers for the package?
