### 2024-08-22 7f283c7
Switched test_geometry over to pytest fully.
