+++ tests/test_geometry.py
@@ -1,5 +1,6 @@
 # Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
 from unittest import TestCase
+import pytest
 
 import montepy
 from montepy.geometry_operators import Operator
@@ -7,437 +8,454 @@ from montepy.input_parser import syntax_node
 from montepy.surfaces.half_space import HalfSpace, UnitHalfSpace
 
 
-class TestHalfSpaceUnit(TestCase):
-    def test_init(self):
-        surface = montepy.surfaces.CylinderOnAxis()
-        node = montepy.input_parser.syntax_node.GeometryTree("hi", {}, "*", " ", " ")
-        half_space = HalfSpace(+surface, Operator.UNION, -surface, node)
-        self.assertIs(half_space.operator, Operator.UNION)
-        self.assertEqual(half_space.left, +surface)
-        self.assertEqual(half_space.right, -surface)
-        self.assertIs(half_space.node, node)
-        with self.assertRaises(TypeError):
-            HalfSpace(surface, Operator.UNION)
-        with self.assertRaises(TypeError):
-            HalfSpace(+surface, "hi")
-        with self.assertRaises(TypeError):
-            HalfSpace(+surface, Operator.UNION, surface)
-        with self.assertRaises(TypeError):
-            HalfSpace(+surface, Operator.UNION, -surface, "hi")
-        with self.assertRaises(ValueError):
-            HalfSpace(+surface, Operator.UNION)
-        with self.assertRaises(ValueError):
-            HalfSpace(+surface, Operator.INTERSECTION)
-
-    def test_get_leaves(self):
-        surface = montepy.surfaces.CylinderOnAxis()
-        cell = montepy.Cell()
-        half_space = -surface & ~cell
-        cells, surfaces = half_space._get_leaf_objects()
-        self.assertEqual(cells, {cell})
-        self.assertEqual(surfaces, {surface})
-
-    def test_len(self):
-        surface = montepy.surfaces.CylinderOnAxis()
-        cell = montepy.Cell()
-        half_space = -surface & ~cell
-        self.assertEqual(len(half_space), 2)
-
-    def test_eq(self):
-        cell1 = montepy.Cell()
-        cell2 = montepy.Cell()
-        half1 = ~cell1 & ~cell2
-        self.assertTrue(half1 == half1)
-        half2 = ~cell1 | ~cell2
-        self.assertTrue(half1 != half2)
-        half2 = ~cell1 & ~cell1
-        self.assertTrue(half1 != half2)
-        half2 = ~cell2 & ~cell2
-        self.assertTrue(half1 != half2)
-        half2 = ~half1
-        half2._operator = Operator.INTERSECTION
-        self.assertTrue(half1 != half2)
-        with self.assertRaises(TypeError):
-            half1 == "hi"
-
-    def test_str(self):
-        cell1 = montepy.Cell()
-        cell1.number = 1
-        cell2 = montepy.Cell()
-        cell2.number = 2
-        half_space = ~cell1 | ~cell2
-        self.assertEqual(str(half_space), "(#1:#2)")
-        half_space = ~cell1 & ~cell2
-        self.assertEqual(str(half_space), "(#1*#2)")
-        half_space = ~(~cell1 & ~cell2)
-        self.assertEqual(str(half_space), "#(#1*#2)")
-        # test that no errors occur
-        repr(half_space)
+def test_init():
+    surface = montepy.surfaces.CylinderOnAxis()
+    node = montepy.input_parser.syntax_node.GeometryTree("hi", {}, "*", " ", " ")
+    half_space = HalfSpace(+surface, Operator.UNION, -surface, node)
+    assert half_space.operator is Operator.UNION
+    assert half_space.left == +surface
+    assert half_space.right == -surface
[…]
