### 2024-08-22 7713327
Updated test to handle having a grouping operator.
