### 2024-08-22 288da57
Tested added parens to scratch logic creation.
