### 2024-08-22 d824cfd
Skipped parens in tree for str.
