### 2024-08-22 4e9599e
Added tests for generating parens.
