### 2024-08-22 cb08869
I can conjugate.
