+++ doc/source/changelog.rst
@@ -12,7 +12,7 @@ MontePy Changelog
 
 * Fixed bug that didn't show metastable states for pretty printing and isotope. Also handled the case that Am-241 metstable states break convention (:issue:`486`).
 * Fixed bug where cell modifiers could be made irrelevant by being added after a comment (:issue:`483`).
-* Fixed bug where parentheses in cell geometry is not properly exported (:pull:`491`).
+* Fixed bug where parentheses in cell geometry are not properly exported (:pull:`491`).
 
 
 0.3.3
