### 2024-08-22 46eae03
fixed list comprehension that deleted comments. Whoops.
