### 2024-08-22 224593e
Upgraded black to match GHA.
