### 2024-08-22 6f0f82e
Updating changelog to next version: 0.4.0
