### 2024-08-22 21b99f3
Update example to use write_problem()
