+++ README.md
@@ -37,8 +37,9 @@ and the Python API documentation.
 * Can parse all surface types except macrobody facets ([Issue #354](https://github.com/idaholab/MontePy/issues/354)).
 * Can read and write out all other MCNP inputs, even if it doesn't not understand them	
 * Attempts to write out the MCNP problem verbatim, even matching the original user formatting. (See [Issues](https://github.com/idaholab/MontePy/issues).)
-* Can quickly access cells, surfaces, and materials by their numbers. For example: `cell = problem.cells[105]`.
-* Can quickly update cell parameters, such as importances. For example `cell.importance.neutron = 2.0`.
+* Can quickly [access cells, surfaces, and materials by their numbers](https://www.montepy.org/starting.html#collections-are-accessible-by-number). For example: `cell = problem.cells[105]`.
+* Can quickly update cell parameters, [such as importances](https://www.montepy.org/starting.html#setting-cell-importances). For example `cell.importance.neutron = 2.0`.
+* Can easily [create universes, and fill other cells with universes](https://www.montepy.org/starting.html#universes).
 * Currently has over 430 test cases.
 
  
