### 2024-08-23 8caf905
Updated example problem to show more features.
