+++ tests/test_material.py
@@ -1,5 +1,4 @@
 # Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
-from unittest import TestCase
 import pytest
 
 import montepy
@@ -14,57 +13,61 @@ from montepy.input_parser.block_type import BlockType
 from montepy.input_parser.mcnp_input import Input
 
 
-class testMaterialClass(TestCase):
-    def test_material_parameter_parsing(self):
-        for line in ["M20 1001.80c 1.0 gas=0", "M20 1001.80c 1.0 gas = 0 nlib = 00c"]:
-            input = Input([line], BlockType.CELL)
-            material = Material(input)
+# test material
+def test_material_parameter_parsing():
+    for line in ["M20 1001.80c 1.0 gas=0", "M20 1001.80c 1.0 gas = 0 nlib = 00c"]:
+        input = Input([line], BlockType.CELL)
+        material = Material(input)
 
-    def test_material_validator(self):
-        material = Material()
-        with self.assertRaises(montepy.errors.IllegalState):
-            material.validate()
-        with self.assertRaises(montepy.errors.IllegalState):
-            material.format_for_mcnp_input((6, 2, 0))
 
-    def test_material_setter(self):
-        in_str = "M20 1001.80c 0.5 8016.80c 0.5"
-        input_card = Input([in_str], BlockType.DATA)
-        material = Material(input_card)
-        material.number = 30
-        self.assertEqual(material.number, 30)
-        with self.assertRaises(TypeError):
-            material.number = "foo"
-        with self.assertRaises(ValueError):
-            material.number = -5
+def test_material_validator():
+    material = Material()
+    with pytest.raises(montepy.errors.IllegalState):
+        material.validate()
+    with pytest.raises(montepy.errors.IllegalState):
+        material.format_for_mcnp_input((6, 2, 0))
 
-    def test_material_str(self):
-        in_str = "M20 1001.80c 0.5 8016.80c 0.4 94239.80c 0.1"
-        input_card = Input([in_str], BlockType.DATA)
-        material = Material(input_card)
-        answers = """\
+
+def test_material_setter():
+    in_str = "M20 1001.80c 0.5 8016.80c 0.5"
+    input_card = Input([in_str], BlockType.DATA)
+    material = Material(input_card)
+    material.number = 30
+    assert material.number == 30
+    with pytest.raises(TypeError):
+        material.number = "foo"
+    with pytest.raises(ValueError):
+        material.number = -5
+
+
+def test_material_str():
+    in_str = "M20 1001.80c 0.5 8016.80c 0.4 94239.80c 0.1"
+    input_card = Input([in_str], BlockType.DATA)
+    material = Material(input_card)
+    answers = """\
 MATERIAL: 20 fractions: atom
- H-1     (80c) 0.5
- O-16    (80c) 0.4
+H-1     (80c) 0.5
+O-16    (80c) 0.4
 Pu-239   (80c) 0.1
 """
-        output = repr(material)
-        print(output)
-        assert output == answers
-        output = str(material)
-        assert output == "MATERIAL: 20, ['hydrogen', 'oxygen', 'plutonium']"
+    output = repr(material)
+    print(output)
+    assert output == answers
+    output = str(material)
+    assert output == "MATERIAL: 20, ['hydrogen', 'oxygen', 'plutonium']"
 
-    def test_material_sort(self):
-        in_str = "M20 1001.80c 0.5 8016.80c 0.5"
-        input_card = Input([in_str], BlockType.DATA)
-        material1 = Material(input_card)
-        in_str = "M30 1001.80c 0.5 8016.80c 0.5"
[…]
