### 2024-08-23 11bbd37
Harmonized isotope with upstream changes to ensure default values are set and handled.
