### 2024-08-23 7eecfe1
Updated isotope str for library being a wrapper object.
