+++ montepy/data_inputs/isotope.py
@@ -254,11 +254,11 @@ class Isotope:
         :returns: a string that can be used in MCNP
         :rtype: str
         """
-        return f"{self.ZAID}.{self.library}" if self.library else self.ZAID
+        return f"{self.ZAID}.{self.library}" if str(self.library) else self.ZAID
 
     def nuclide_str(self):
         meta_suffix = f"m{self.meta_state}" if self.is_metastable else ""
-        suffix = f".{self._library}" if self._library else ""
+        suffix = f".{self._library}" if str(self._library) else ""
         return f"{self.element.symbol}-{self.A}{meta_suffix}{suffix}"
 
     def get_base_zaid(self):
@@ -372,7 +372,7 @@ class Isotope:
 
     def __str__(self):
         meta_suffix = f"m{self.meta_state}" if self.is_metastable else ""
-        suffix = f" ({self._library})" if self._library else ""
+        suffix = f" ({self._library})" if str(self._library) else ""
         return f"{self.element.symbol:>2}-{self.A:<3}{meta_suffix:<2}{suffix}"
 
     def __hash__(self):
