### 2024-08-23 dcb50f3
Started implement isotope slicing.
