### 2024-08-23 2cc9750
Started test get/setitem for material.
