### 2024-08-23 360c2d5
Updated fancy_name to accept just an element outside a tuple.
