### 2024-08-23 3763a52
Made isotope library use meta-programming.
