### 2024-08-23 78090fe
Ensure _is_atom_franction always set.
