+++ montepy/data_inputs/material.py
@@ -40,6 +40,7 @@ class Material(data_input.DataInputAbstract, Numbered_MCNP_Object):
         self._material_components = {}
         self._pointers = co.defaultdict(lambda: co.defaultdict(dict))
         self._thermal_scattering = None
+        self._is_atom_fraction = False
         self._number = self._generate_default_node(int, -1)
         super().__init__(input)
         if input:
