### 2024-08-23 d88ceeb
Fixing bugs with get/setitem.
