### 2024-08-23 f7567fc
Avoided trying to look for non-existant isotopes.
