+++ tests/test_material.py
@@ -275,25 +275,25 @@ def test_isotope_str():
 @pytest.mark.parametrize(
     "input, Z, A, meta, library",
     [
-        (1001, 1, 1, None, ""),
-        ("1001.80c", 1, 1, None, "80c"),
-        ("h1", 1, 1, None, ""),
-        ("h-1", 1, 1, None, ""),
-        ("h", 1, 0, None, ""),
-        ("hydrogen-1", 1, 1, None, ""),
-        ("hydrogen", 1, 0, None, ""),
-        ("hydrogen1", 1, 1, None, ""),
+        (1001, 1, 1, 0, ""),
+        ("1001.80c", 1, 1, 0, "80c"),
+        ("h1", 1, 1, 0, ""),
+        ("h-1", 1, 1, 0, ""),
+        ("h", 1, 0, 0, ""),
+        ("hydrogen-1", 1, 1, 0, ""),
+        ("hydrogen", 1, 0, 0, ""),
+        ("hydrogen1", 1, 1, 0, ""),
         ("hydrogen1m3", 1, 1, 3, ""),
         ("hydrogen1m3.80c", 1, 1, 3, "80c"),
         ("92635m2.710nc", 92, 235, 3, "710nc"),
-        (Isotope("1001.80c"), 1, 1, None, "80c"),
+        (Isotope("1001.80c"), 1, 1, 0, "80c"),
         ((92, 235, 1, "80c"), 92, 235, 1, "80c"),
         ((Element(92), 235, 1, "80c"), 92, 235, 1, "80c"),
-        ((Element(92), 235), 92, 235, None, ""),
-        (("U", 235), 92, 235, None, ""),
-        ((92, 235), 92, 235, None, ""),
-        (("uRanium", 235), 92, 235, None, ""),
-        ((Element(92),), 92, 0, None, ""),
+        ((Element(92), 235), 92, 235, 0, ""),
+        (("U", 235), 92, 235, 0, ""),
+        ((92, 235), 92, 235, 0, ""),
+        (("uRanium", 235), 92, 235, 0, ""),
+        ((Element(92),), 92, 0, 0, ""),
     ],
 )
 def test_fancy_names(input, Z, A, meta, library):
@@ -314,7 +314,10 @@ def big_material():
         "h2",
         "h3",
         "th232",
+        "th232.701nc",
         "U235",
+        "U235.80c",
+        "U235m1.80c",
         "u238",
         "am242",
         "am242m1",
@@ -331,7 +334,7 @@ def big_material():
     "index",
     [
         1001,
-        "1001.80c",
+        "1001.00c",
         "h1",
         "h-1",
         "h",
@@ -339,8 +342,9 @@ def big_material():
         "hydrogen",
         "hydrogen1",
         "hydrogen1m3",
-        "hydrogen1m3.80c",
-        "92635m2.710nc",
+        "hydrogen1m3.00c",
+        "Th232.710nc",
+        "92635",
         (Isotope("1001.80c"),),
         (92, 235, 1, "80c"),
         (Element(92), 235, 1, "80c"),
