### 2024-08-23 4910729
Removed magic number from Isotopes.
