### 2024-08-23 43bbdd2
Changed default meta_state to 0 from None.
