### 2024-08-23 1048375
Allowed passing just an element for slicing.
