### 2024-08-25 73a9b1c
Started changelog for material redesign.
