+++ doc/source/migrate0_1.rst
@@ -1,0 +1,12 @@
+.. _migrate01:
+
+Migration plan for MontePy 1.0.0
+================================
+
+.. meta::
+   :description: Foo bar
+
+Necessity
+---------
+
+See :issue:`504`.
