### 2024-08-26 873b0d0
Added deprecation not to material_component.
