+++ montepy/data_inputs/material_component.py
@@ -15,6 +15,11 @@ class MaterialComponent:
 
     For example: this may be H-1 in water: like 1001.80c — 0.6667
 
+    .. deprecated:: 0.4.1
+        MaterialComponent has been deprecated as part of a redesign for the material
+        interface due to a critical bug in how MontePy handles duplicate nuclides.
+        See `migrate01`_.
+
     :param isotope: the Isotope object representing this isotope
     :type isotope: Isotope
     :param fraction: the fraction of this component in the material
