### 2024-08-26 #205 @kellyrowland
hello @jpmorgan98 @gwenchee 👋 would you be interested in and available to review this pyOpenSci submission?

the reviewer template that you would use can be seen at https://www.pyopensci.org/software-peer-review/appendices/templates.html#peer-review-template .
