### 2024-08-26 idaholab/MontePy#433 @MicahGale
https://github.com/kordusas/npp_mcnp_plugin
