### 2024-08-27 7b3c57f
Squashed profile commits that mat_optimize had
