### 2024-08-27 ef354fd
Tests now need to be able to build package.
