### 2024-08-27 629e531
Made a big annoying model to stress MontePy for profiling and benchmarking.
