### 2024-08-27 0b81466
Increased failure threshold to 4 minutes.
