### 2024-08-27 35c8e64
Learned how to actually get the results from a profile.
