### 2024-08-28 f1a2118
Made mechanism for data inputs to flag self for deletion.
