### 2024-08-28 7ffa073
Moved material linking to thermal_scattering with hashing.
