+++ montepy/data_inputs/thermal_scattering.py
@@ -122,18 +122,39 @@ class ThermalScatteringLaw(DataInputAbstract):
 
         :param data_inputs: a list of the data inputs in the problem
         :type data_inputs: list
+        :returns: True iff this input should be removed from ``problem.data_inputs``
+        :rtype: bool
         """
+        # use caching first
+        if self._problem:
+            try:
+                mat = self._problem.materials[self.old_number]
+            except KeyError:
+                raise MalformedInputError(
+                    self._input, "MT input is detached from a parent material"
+                )
+        # brute force it
         found = False
-        for input in data_inputs:
-            if isinstance(input, montepy.data_inputs.material.Material):
-                if input.number == self.old_number:
+        for data_input in data_inputs:
+            if isinstance(data_input, montepy.data_inputs.material.Material):
+                if data_input.number == self.old_number:
+                    mat = data_input
                     found = True
-                    self._parent_material = input
-
+                    break
+        # actually update things
         if not found:
             raise MalformedInputError(
                 self._input, "MT input is detached from a parent material"
             )
+
+        if mat.thermal_scattering:
+            raise MalformedInputError(
+                self,
+                f"Multiple MT inputs were specified for this material: {self.old_number}.",
+            )
+        mat.thermal_scattering = self
+        self._parent_material = mat
+        return True
 
     def __str__(self):
         return f"THERMAL SCATTER: {self.thermal_scattering_laws}"
