+++ benchmark/benchmark_big_model.py
@@ -3,7 +3,7 @@ import montepy
 import time
 import tracemalloc
 
-FAIL_THRESHOLD = 500
+FAIL_THRESHOLD = 30
 
 tracemalloc.start()
 start = time.time()
