### 2024-08-28 884bba7
Add profiling and benckmarking (#513)
* Started profiling test suite.

* Tests now need to be able to build package.

* Filtered less for montepy

* Made a big annoying model to stress MontePy for profiling and benchmarking.

* Enabled big model CI work.

* Fixed bad path.

* Fixed benchmark and gave failure condition.

* Increased failure threshold to 4 minutes.

* Learned how to actually get the results from a profile.

* GHA is slow: increased threshold.

* Updated benchmark to not be near line lenght max.
