### 2024-08-28 9ab65f6
Implemented reading input from a stream.
