### 2024-08-28 97683b9
Test read/write all input files.
