### 2024-08-28 94233d4
Dumped whole file to stdout.
