+++ tests/test_integration.py
@@ -484,7 +484,7 @@ def test_importance_format_unmutated(importance_problem):
     print(output)
     assert len(output) == 2
     assert "imp:n,p 1 1 1 0 3" == output[0]
-    assert "imp:e   0 2R 1 2" == output[1]
+    assert "imp:e   0 2r 1 r" == output[1]
 
 
 def test_importance_format_mutated(importance_problem):
@@ -1073,15 +1073,18 @@ def test_alternate_encoding():
 
 @pytest.mark.parametrize(
     "file",
-    set((Path("tests") / "inputs").iterdir())
-    - {
-        Path("tests") / "inputs" / p
-        for p in constants.BAD_INPUTS | constants.IGNORE_FILES
-    },
+    {"tests/inputs/test_importance.imcnp"},
+    # set((Path("tests") / "inputs").iterdir())
+    # - {
+    #    Path("tests") / "inputs" / p
+    #    for p in constants.BAD_INPUTS | constants.IGNORE_FILES
+    # },
 )
 def test_read_write_cycle(file):
     print(f"Testing against {file} *********************")
     problem = montepy.read_input(file)
+    print(problem.cells._importance._real_tree[montepy.particle.Particle.ELECTRON])
+    print()
     fh = io.StringIO()
     problem.write_problem(fh)
     fh.seek(0)
