### 2024-08-28 c577a88
Fixed logic to detect when you can smash repeats together.
