### 2024-08-28 c4b56ac
Updated logic to test for matching value with valueNode.
