### 2024-08-28 4284c03
Updated logic to always check for edge node matching.
