+++ tests/test_syntax_parsing.py
@@ -823,38 +823,6 @@ class TestShortcutNode(TestCase):
         with self.assertRaises(TypeError):
             short.end_padding = " "
 
-    """
-    Most examples, unless otherwise noted are taken from Section 2.8.1
-    of LA-UR-17-29981.
-    """
-    tests = {
-        "1 3M 2r": [1, 3, 3, 3],
-        # unofficial
-        "0.01 2ILOG 10": [0.01, 0.1, 1, 10],
-        "1 3M I 4": [1, 3, 3.5, 4],
-        "1 3M 3M": [1, 3, 9],
-        "1 2R 2I 2.5": [1, 1, 1, 1.5, 2, 2.5],
-        "1 R 2m": [1, 1, 2],
-        "1 R R": [1, 1, 1],
-        "1 2i 4 3m": [1, 2, 3, 4, 12],
-        # unofficial
-        "1 i 3": [1, 2, 3],
-        # unofficial
-        "1 ilog 100": [1, 10, 100],
-        # last official one
-        "1 2i 4 2i 10": [
-            1,
-            2,
-            3,
-            4,
-            6,
-            8,
-            10,
-        ],
-        "1 2j 4": [1, montepy.Jump(), montepy.Jump(), 4],
-        "1 j": [1, montepy.Jump()],
-    }
-
     def test_shortcut_expansion(self):
         invalid = [
             "3J 4R",
@@ -869,7 +837,7 @@ class TestShortcutNode(TestCase):
         ]
 
         parser = ShortcutTestFixture()
-        for test, answer in self.tests.items():
+        for test, answer in tests.items():
             print(test)
             input = Input([test], BlockType.DATA)
             parsed = parser.parse(input.tokenize())
@@ -908,35 +876,76 @@ class TestShortcutNode(TestCase):
             for val, gold in zip(parsed, answer):
                 self.assertAlmostEqual(val.value, gold)
 
-    def test_shortcut_format(self):
-        parser = ShortcutTestFixture()
-        for in_str, answer in [
-            ("1 5R", "1 5R"),
-            ("1 5r", "1 5r"),
-            ("1 r", "1 r"),
-            ("1 2i 5", "1 2i 5"),
-            ("1 2ilog 5", "1 2ilog 5"),
-            ("1 r 2i 5", "1 r 2i 5"),
-            ("1 r 2ilog 5", "1 r 2ilog 5"),
-            ("1 5M", "1 5M"),
-            ("1 5m", "1 5m"),
-            ("2j ", "2j "),
-            ("2j", "2j"),
-            ("2J ", "2J "),
-            ("J", "J"),
-        ]:
-            print(in_str, answer)
-            input = Input([in_str], BlockType.CELL)
-            shortcut = parser.parse(input.tokenize())
-            self.assertEqual(shortcut.format(), answer)
-        # try jump with empty jump shortcut
-        shortcut.nodes.clear()
-        self.assertEqual(shortcut.format(), "")
-        for in_str in self.tests.keys():
-            print(in_str)
-            input = Input([in_str], BlockType.CELL)
-            shortcut = parser.parse(input.tokenize())
-            self.assertEqual(shortcut.format(), in_str)
+
+"""
+Most examples, unless otherwise noted are taken from Section 2.8.1
+of LA-UR-17-29981.
+"""
+tests = {
+    "1 3M 2r": [1, 3, 3, 3],
+    # unofficial
[…]
