### 2024-08-29 1529886
Made logic clearer and more pythonic.
