### 2024-08-29 8ff6b69
Made it possible to have integer interpolations.
