### 2024-08-29 73800b8
Started framework for geometry tree recompressing interpolations.
