### 2024-08-29 b8b8b3b
Switched to preserving shortcut type for geometry stuff.
