### 2024-08-29 0582457
Exposed shortcut type.
