### 2024-08-29 dd16501
Made system to make a shortcutNode from scratch kind of.
