### 2024-08-29 e9d2e4e
Formatted shortcuts inside of GeometryTree
