### 2024-08-29 4902c43
Wrote up a description of the deprecations.
