### 2024-08-29 0e9f406
Avoided triggering deprecation warnings internally.
