### 2024-08-29 8e88ce3
Added #506 to changelog.
