### 2024-08-30 fabfb52
Made system to ignore default parameters for trailing comments.
