### 2024-08-30 f033d97
Avoid cross-block trailing comment transfer.
