### 2024-08-30 3edca07
Updated delete trailing comment to match new get_trailing_comment.
