### 2024-08-30 08a0c98
Skip empty nodes in the tail.
