### 2024-08-31 5029946
Updated get trailing comment logic to handle valueNodes in tree.
