### 2024-08-31 f87c875
Updated complex logic for deleteing trailing comments.
