### 2024-08-31 ee08583
Updated geometry expansion tests to be more exhaustive.
