### 2024-08-31 3f7f277
fixed bugs in M in geometry.
