### 2024-08-31 a00afe4
Added more edge cases.
