### 2024-09-02 97bca51
Made system for tracking left and right shortuts
