### 2024-09-02 24ca152
Updated shortcut finding logic to find shortcuts in the middle.
