+++ tests/test_syntax_parsing.py
@@ -865,8 +865,8 @@ class TestShortcutNode(TestCase):
         ("1 2i 4 ", [1, 2, 3, 4], None),
         ("1 2i 4 ", [1, 2, 3, 4], None),
         ("1 1 2i 4 ", [1, 1, 2, 3, 4], None),
-        ("1 1 2i 4 5 6 ", [1, 1, 2, 3, 4, 5, 6], None),
-        ("1 1 2i 4:5 6 ", [1, 1, 2, 3, 4, 5, 6], None),
+        ("1 1 2i 4 5 6 ", [1, 1, 2, 3, 4, 5, 6], "1 1 2I 4  5 6"),
+        ("1 1 2i 4:5 6 ", [1, 1, 2, 3, 4, 5, 6], "1 1 2I 4 :5 6"),
         ("1 ilog 100 ", [1, 10, 100], "1 1ILOG 100"),
         # secretly test iterator
         ("#1", [1], None),
