### 2024-09-02 aca2f85
Updated changelog with #520 PR.
