### 2024-09-02 b6ec5d4
Fixed over-appending leaves to the new list.
