+++ doc/source/users.rst
@@ -9,6 +9,6 @@ User Guide
    utilities
    tricks
    faq
-   migrate0_1
+   migrations/migrate0_1
    changelog
    publications
