### 2024-09-02 811ebed
Marked Isotope class as deprecated.
