+++ tests/test_material.py
@@ -87,11 +87,16 @@ def test_material_format_mcnp():
 )
 def test_material_comp_init(isotope, conc, error):
     with pytest.raises(error):
-        MaterialComponent(Isotope(isotope), conc)
+        MaterialComponent(Isotope(isotope, suppress_warning=True), conc, True)
+
+
+def test_mat_comp_init_warn():
+    with pytest.warns(DeprecationWarning):
+        MaterialComponent(Isotope("1001.80c", suppress_warning=True), 0.1)
 
 
 def test_material_comp_fraction_setter():
-    comp = MaterialComponent(Isotope("1001.80c"), 0.1)
+    comp = MaterialComponent(Isotope("1001.80c", suppress_warning=True), 0.1, True)
     comp.fraction = 5.0
     assert comp.fraction == pytest.approx(5.0)
     with pytest.raises(ValueError):
@@ -101,7 +106,7 @@ def test_material_comp_fraction_setter():
 
 
 def test_material_comp_fraction_str():
-    comp = MaterialComponent(Isotope("1001.80c"), 0.1)
+    comp = MaterialComponent(Isotope("1001.80c", suppress_warning=True), 0.1, True)
     str(comp)
     repr(comp)
 
@@ -115,9 +120,9 @@ def test_material_update_format():
     print(material.format_for_mcnp_input((6, 2, 0)))
     assert "8016" in material.format_for_mcnp_input((6, 2, 0))[0]
     # addition
-    isotope = Isotope("2004.80c")
+    isotope = Isotope("2004.80c", suppress_warning=True)
     with pytest.deprecated_call():
-        material.material_components[isotope] = MaterialComponent(isotope, 0.1)
+        material.material_components[isotope] = MaterialComponent(isotope, 0.1, True)
         print(material.format_for_mcnp_input((6, 2, 0)))
         assert "2004" in material.format_for_mcnp_input((6, 2, 0))[0]
         # update
@@ -126,7 +131,7 @@ def test_material_update_format():
         material.material_components[isotope].fraction = 0.7
         print(material.format_for_mcnp_input((6, 2, 0)))
         assert "0.7" in material.format_for_mcnp_input((6, 2, 0))[0]
-        material.material_components[isotope] = MaterialComponent(isotope, 0.6)
+        material.material_components[isotope] = MaterialComponent(isotope, 0.6, True)
         print(material.format_for_mcnp_input((6, 2, 0)))
         assert "0.6" in material.format_for_mcnp_input((6, 2, 0))[0]
         # delete
@@ -177,28 +182,29 @@ def test_bad_init(line):
 
 class TestIsotope(TestCase):
     def test_isotope_init(self):
-        isotope = Isotope("1001.80c")
+        with pytest.warns(FutureWarning):
+            isotope = Isotope("1001.80c")
         self.assertEqual(isotope.ZAID, "1001")
         self.assertEqual(isotope.Z, 1)
         self.assertEqual(isotope.A, 1)
         self.assertEqual(isotope.element.Z, 1)
         self.assertEqual(isotope.library, "80c")
         with self.assertRaises(ValueError):
-            Isotope("1001.80c.5")
+            Isotope("1001.80c.5", suppress_warning=True)
         with self.assertRaises(ValueError):
-            Isotope("hi.80c")
+            Isotope("hi.80c", suppress_warning=True)
 
     def test_isotope_metastable_init(self):
-        isotope = Isotope("13426.02c")
+        isotope = Isotope("13426.02c", suppress_warning=True)
         self.assertEqual(isotope.ZAID, "13426")
         self.assertEqual(isotope.Z, 13)
         self.assertEqual(isotope.A, 26)
[…]
