### 2024-09-03 b42f4a8
Added tests for mark_last_leaf_shortcut.
