### 2024-09-03 847d2d2
Started testing for flatten_shortcut.
