+++ tests/test_syntax_parsing.py
@@ -913,7 +913,39 @@ def test_shortcut_geometry_expansion(test, answer, form_ans):
     print(test)
     input = Input([test], BlockType.CELL)
     parsed = parser.parse(input.tokenize())
-    print(parsed)
+    for val, gold in zip(parsed, answer):
+        assert val.value == gold
+    if form_ans:
+        assert parsed.format().rstrip() == form_ans
+    else:
+        assert parsed.format().rstrip() == test.upper().rstrip()
+
+
+@pytest.mark.parametrize(
+    "test, answer, form_ans",
+    [
+        ("1 3r ", [1, 1, 1, 1], None),
+        ("1 1 3r ", [1, 1, 1, 1, 1], None),
+        ("1 1 2M 3r ", [1, 1, 2, 2, 2, 2], None),
+        ("1 -2M ", [1, -2], None),
+        ("1 2i 4 ", [1, 2, 3, 4], None),
+        ("1 2i 4 ", [1, 2, 3, 4], None),
+        ("1 1 2i 4 ", [1, 1, 2, 3, 4], None),
+        ("1 1 2i 4 5 6 ", [1, 1, 2, 3, 4, 5, 6], "1 1 2I 4  5 6"),
+        ("1 1 2i 4:5 6 ", [1, 1, 2, 3, 4, 5, 6], "1 1 2I 4 :5 6"),
+        ("1 ilog 100 ", [1, 10, 100], "1 1ILOG 100"),
+        # secretly test iterator
+        ("#1", [1], None),
+        ("#(1 2 3)", [1, 2, 3], None),
+        ("1 2:( 3 4 5)", [1, 2, 3, 4, 5], None),
+    ],
+)
+def test_shortcut_flatten(test, length, indices):
+
+    parser = ShortcutGeometryTestFixture()
+    print(test)
+    input = Input([test], BlockType.CELL)
+    parsed = parser.parse(input.tokenize())
     for val, gold in zip(parsed, answer):
         assert val.value == gold
     if form_ans:
