### 2024-09-03 06a44cb
Finished test for flatten_shortcuts.
