+++ tests/test_syntax_parsing.py
@@ -922,22 +922,22 @@ def test_shortcut_geometry_expansion(test, answer, form_ans):
 
 
 @pytest.mark.parametrize(
-    "test, answer, form_ans",
+    "test, length, indices",
     [
-        ("1 3r ", [1, 1, 1, 1], None),
-        ("1 1 3r ", [1, 1, 1, 1, 1], None),
-        ("1 1 2M 3r ", [1, 1, 2, 2, 2, 2], None),
-        ("1 -2M ", [1, -2], None),
-        ("1 2i 4 ", [1, 2, 3, 4], None),
-        ("1 2i 4 ", [1, 2, 3, 4], None),
-        ("1 1 2i 4 ", [1, 1, 2, 3, 4], None),
-        ("1 1 2i 4 5 6 ", [1, 1, 2, 3, 4, 5, 6], "1 1 2I 4  5 6"),
-        ("1 1 2i 4:5 6 ", [1, 1, 2, 3, 4, 5, 6], "1 1 2I 4 :5 6"),
-        ("1 ilog 100 ", [1, 10, 100], "1 1ILOG 100"),
+        ("1 3r ", 2, {1: Shortcuts.REPEAT}),
+        ("1 1 3r ", 3, {2: Shortcuts.REPEAT}),
+        ("1 1 2M 3r ", 4, {2: Shortcuts.MULTIPLY, 3: Shortcuts.REPEAT}),
+        ("1 -2M ", 2, {1: Shortcuts.MULTIPLY}),
+        ("1 2i 4 ", 1, {0: Shortcuts.INTERPOLATE}),
+        ("1 2i 4 ", 1, {0: Shortcuts.INTERPOLATE}),
+        ("1 1 2i 4 ", 2, {1: Shortcuts.INTERPOLATE}),
+        ("1 1 2i 4 5 6 ", 4, {1: Shortcuts.INTERPOLATE}),
+        ("1 1 2i 4:5 6 ", 4, {1: Shortcuts.INTERPOLATE}),
+        ("1 ilog 100 ", 1, {0: Shortcuts.INTERPOLATE}),
         # secretly test iterator
-        ("#1", [1], None),
-        ("#(1 2 3)", [1, 2, 3], None),
-        ("1 2:( 3 4 5)", [1, 2, 3, 4, 5], None),
+        ("#1", 1, {}),
+        ("#(1 2 3)", 3, {}),
+        ("1 2:( 3 4 5)", 5, {}),
     ],
 )
 def test_shortcut_flatten(test, length, indices):
@@ -946,12 +946,11 @@ def test_shortcut_flatten(test, length, indices):
     print(test)
     input = Input([test], BlockType.CELL)
     parsed = parser.parse(input.tokenize())
-    for val, gold in zip(parsed, answer):
-        assert val.value == gold
-    if form_ans:
-        assert parsed.format().rstrip() == form_ans
-    else:
-        assert parsed.format().rstrip() == test.upper().rstrip()
+    flatpack = parsed._flatten_shortcut()
+    assert len(flatpack) == length
+    for index, short_type in indices.items():
+        assert isinstance(flatpack[index], ShortcutNode)
+        assert flatpack[index].type == short_type
 
 
 """
