### 2024-09-03 058b757
Cleaned up with @tjlaboss feedback.
