### 2024-09-03 f7ff8ba
Added test replicate new edge case from @tjlaboss.
