### 2024-09-03 9483f89
Added more exhaustive shortcut tests.
