### 2024-09-03 6b6cdcb
Expanded using last node on formatting to all shortcuts, and handled interpolate messiness.
