+++ tests/test_syntax_parsing.py
@@ -897,6 +897,7 @@ def test_shortcut_expansion_valid(test, answer):
     print(test)
     input = Input([test], BlockType.DATA)
     parsed = parser.parse(input.tokenize())
+    print(parsed)
     for val, gold in zip(parsed, answer):
         if val.value is None:
             assert gold == montepy.Jump()
@@ -942,7 +943,7 @@ def test_shortcut_expansion_invalid(test):
         ("1 1 2i 4 5 6 ", [1, 1, 2, 3, 4, 5, 6], "1 1 2I 4  5 6"),
         ("1 1 2i 4:5 6 ", [1, 1, 2, 3, 4, 5, 6], "1 1 2I 4 :5 6"),
         ("1 ilog 100 ", [1, 10, 100], "1 1ILOG 100"),
-        ("1 1r ilog 100 ", [1, 1, 10, 100], "1 1r 1ILOG 100"),
+        ("1 1r ilog 100 ", [1, 1, 10, 100], "1 1R 1ILOG 100"),
         # secretly test iterator
         ("#1", [1], None),
         ("#(1 2 3)", [1, 2, 3], None),
