### 2024-09-03 9430561
Updated logic for multiply shortcuts using neigbors.
