### 2024-09-03 5a8b34a
Created system to track starting value for interpolate in geometry trees.
