### 2024-09-03 628cd9d
Added doc strings for new functions.
