### 2024-09-04 bd3e748
Deleted all previously deprecated code.
