+++ montepy/data_inputs/cell_modifier.py
@@ -266,40 +266,3 @@ class CellModifierInput(DataInputAbstract):
             self._update_values()
             return self.wrap_string_for_mcnp(self._format_tree(), mcnp_version, True)
         return []
-
-    @property
-    def has_changed_print_style(self):  # pragma: no cover
-        """
-        returns true if the printing style for this modifier has changed
-        from cell block to data block, or vice versa.
-
-        .. deprecated:: 0.2.0
-            This property is no longer needed and overly complex.
-
-        :returns: true if the printing style for this modifier has changed
-        :rtype: bool
-        :raises DeprecationWarning: raised always.
-        """
-        warnings.warn(
-            "has_changed_print_style will be removed soon.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        if self._problem:
-            print_in_cell_block = not self._problem.print_in_data_block[
-                self.class_prefix
-            ]
-            set_in_cell_block = print_in_cell_block
-            if not self.in_cell_block:
-                for cell in self._problem.cells:
-                    attr = montepy.Cell._CARDS_TO_PROPERTY[type(self)][0]
-                    modifier = getattr(cell, attr)
-                    if modifier.has_information:
-                        set_in_cell_block = modifier.set_in_cell_block
-                    break
-            else:
-                if self.has_information:
-                    set_in_cell_block = self.set_in_cell_block
-            return print_in_cell_block ^ set_in_cell_block
-        else:
-            return False
