### 2024-09-04 a9da602
Updated changelog with deleted code.
