### 2024-09-05 6e8e0ae
Switched away from mat_comp dict.
