### 2024-09-05 d69a019
Started get/set item, and base dunders.
