+++ montepy/data_inputs/material.py
@@ -112,33 +112,79 @@ class Material(data_input.DataInputAbstract, Numbered_MCNP_Object):
         The keys are :class:`~montepy.data_inputs.isotope.Isotope` instances, and the values are
         :class:`~montepy.data_inputs.material_component.MaterialComponent` instances.
 
-        .. deprecated:: 0.4.0
-            Accessing this dictionary directly is deprecated.
-            Instead access the nuclides directly with the keys.
+        .. deprecated:: 0.4.1
+            MaterialComponent has been deprecated as part of a redesign for the material
+            interface due to a critical bug in how MontePy handles duplicate nuclides.
+            See :ref:`migrate 0 1`.
 
-        :rtype: dict
+        :raises DeprecationWarning: This has been fully deprecated and cannot be used.
         """
-        return self._material_components
+        raise DeprecationWarning(
+            f"""material_components is deprecated, and has been removed in MontePy 1.0.0.
+See <https://www.montepy.org/migrations/migrate0_1.html> for more information """
+        )
 
-    def __getitem__(self, key):
+    def __getitem__(self, idx):
         """ """
-        # decide if this is a slice
-        if isinstance(key, (tuple, Element)):
-            if isinstance(key, Element):
-                return self.__get_slice((key,))
-            # TODO think about upper limit
-            if len(key) <= 3:
-                return self.__get_slice(key)
-            if any([isinstance(s, slice) for s in key]):
-                return self.__get_slice(key)
-        pointer = self.__get_pointer_iso(key)
-        return self.material_components[pointer].fraction
+        if not isinstance(idx, (int, slice)):
+            raise TypeError(f"Not a valid index. {idx} given.")
+        return self._components[idx]
+
+    def __iter__(self):
+        return iter(self._components)
+
+    def __setitem__(self, key, newvalue):
+        """ """
+        if not isinstance(idx, (int, slice)):
+            raise TypeError(f"Not a valid index. {idx} given.")
+        self._check_valid_comp(newvalue)
+        self._components[idx] = newvalue
+
+    def _check_valid_comp(self, newvalue):
+        if not isinstance(newvalue, tuple):
+            raise TypeError(
+                f"Invalid component given. Must be tuple of Isotope, fraction. {newvalue} given."
+            )
+        if len(newvalue) != 2:
+            raise ValueError(
+                f"Invalid component given. Must be tuple of Isotope, fraction. {newvalue} given."
+            )
+        if not isinstance(newvalue[0], Isotope):
+            raise TypeError(f"First element must be an Isotope. {newvalue[0]} given.")
+        if not isinstance(newvalue[1], (float, int)):
+            raise TypeError(
+                f"Second element must be a fraction greater than 0. {newvalue[1]} given."
+            )
+        if newvalue[1] < 0.0:
+            raise TypeError(
+                f"Second element must be a fraction greater than 0. {newvalue[1]} given."
+            )
+
+    def __delitem__(self, idx):
+        if not isinstance(idx, (int, slice)):
+            raise TypeError(f"Not a valid index. {idx} given.")
+        del self._components[idx]
+
+    def append(self, obj):
+        self._check_valid_comp(obj)
+        self._components.append(obj)
+
+    def find(
+        self, fancy_name=None, element=None, A=None, meta_isomer=None, library=None
+    ):
+        """ """
+        pass
+
+    def __iadd__(self, other):
+        pass
+
+    def __bool__(self):
+        pass
+
+    def __add__(self, other):
+        pass
 
     def __get_pointer_iso(self, key):
[…]
