### 2024-09-07 0b6bd38
Update GHA to get coverage for unit tests only.
