### 2024-09-07 457284d
Reduced coverage threshold while dealing with unit testing transition.
