+++ tests/unit/test_main.py
@@ -4,7 +4,7 @@ import montepy
 from montepy import __main__ as main
 from montepy.errors import *
 from unittest import TestCase
-from tests import constants
+from .. import constants
 import os
 
 
