+++ .github/workflows/main.yml
@@ -64,7 +64,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - run: pip install --user . montepy[test]
       - run: pip install --user . montepy[build]
-      - run: coverage run -m pytest tests/unit --junitxml=test_report.xml
+      - run: pytest tests/unit --junitxml=test_report.xml --cov
         name: Collect unit test coverage
       - run: pytest tests/integr
         name: Run integration tests
