### 2024-09-09 1d45c9c
Fixed material parser to be able to parse plib=84p.
