+++ tests/test_numbered_collection.py
@@ -317,11 +317,13 @@ def test_data_append(cp_simple_problem):
     prob.materials.append(new_mat)
     assert new_mat in prob.materials
     assert new_mat in prob.data_inputs
+    assert prob.data_inputs.count(new_mat) == 1
     # trigger getting data_inputs end
     prob.materials.clear()
     prob.materials.append(new_mat)
     assert new_mat in prob.materials
     assert new_mat in prob.data_inputs
+    assert prob.data_inputs.count(new_mat) == 1
     prob.data_inputs.clear()
     prob.materials._last_index = None
     new_mat = copy.deepcopy(next(iter(prob.materials)))
@@ -329,6 +331,7 @@ def test_data_append(cp_simple_problem):
     prob.materials.append(new_mat)
     assert new_mat in prob.materials
     assert new_mat in prob.data_inputs
+    assert prob.data_inputs.count(new_mat) == 1
     # trigger getting index of last material
     prob.materials._last_index = None
     new_mat = copy.deepcopy(next(iter(prob.materials)))
