### 2024-09-10 31cb4eb
Implemented problem weakref.
