### 2024-09-10 057486b
Implemented weakref hiding for pickles.
