### 2024-09-10 e108ac3
Made sure linking operation takes place in the right order.
