### 2024-09-10 128c651
Update gold value with new line.
