### 2024-09-10 9a88b22
Updated test with new input model.
