### 2024-09-10 39a12c4
Tried to figure out where this stupid valuenode comment is coming from.
