### 2024-09-11 9da343a
Setup system to track which particles were paired up.
