### 2024-09-11 5882881
Updated to use actual trees, and ensured all affected trees cleared.
