+++ montepy/data_inputs/importance.py
@@ -360,23 +360,30 @@ class Importance(CellModifierInput):
     def _delete_trailing_comment(self):
         for part, tree in reversed(self._real_tree.items()):
             tree._delete_trailing_comment()
-            part_tree = self._particle_importances[part]
-            part_tree["data"]._delete_trailing_comment()
+            self.__delete_common_trailing(part)
             break
+
+    def __delete_common_trailing(self, part):
+        to_delete = {part}
+        for combo_set in self._part_combos:
+            if part in combo_set:
+                to_delete |= combo_set
+        for part in to_delete:
+            self._particle_importances[part]["data"]._delete_trailing_comment()
 
     def _grab_beginning_comment(self, new_padding):
         last_tree = None
         last_padding = None
-        for tree in self._real_tree.values():
+        for part, tree in self._particle_importances.items():
             if last_padding is not None and last_tree is not None:
                 last_tree._grab_beginning_comment(last_padding)
-                last_tree._delete_trailing_comment()
+                self.__delete_common_trailing(part)
             last_padding = tree.get_trailing_comment()
             last_tree = tree
         if new_padding:
-            next(iter(self._real_tree.values()))["start_pad"]._grab_beginning_comment(
-                new_padding
-            )
+            next(iter(self._particle_importances.values()))[
+                "start_pad"
+            ]._grab_beginning_comment(new_padding)
 
 
 def _generate_default_data_tree(particle):
