### 2024-09-11 a30549b
Fixed bug where the parameters weren't actually loaded into syntax tree.
