### 2024-09-11 08ded2f
Display "a" version under the logo
