### 2024-09-11 27a6ae5
Correct the comments in doc/source/conf.py
