### 2024-09-12 0ed30ef
Refactored Isotope to Nuclide.
