+++ montepy/data_inputs/isotope.py
@@ -1,0 +1,8 @@
+class Isotope:
+
+    def __init__(self, *args, **kwargs):
+        """ """
+        raise DeprecationWarning(
+            "montepy.data_inputs.isotope.Isotope is deprecated and is renamed: Nuclide.\n"
+            "See <https://www.montepy.org/migrations/migrate0_1.html> for more information "
+        )
