+++ tests/test_material.py
@@ -4,7 +4,7 @@ import pytest
 import montepy
 
 from montepy.data_inputs.element import Element
-from montepy.data_inputs.isotope import Isotope, Library
+from montepy.data_inputs.nuclide import Nuclide, Library
 from montepy.data_inputs.material import Material
 from montepy.data_inputs.material_component import MaterialComponent
 from montepy.data_inputs.thermal_scattering import ThermalScatteringLaw
@@ -90,16 +90,16 @@ def test_material_format_mcnp():
 )
 def test_material_comp_init(isotope, conc, error):
     with pytest.raises(error):
-        MaterialComponent(Isotope(isotope, suppress_warning=True), conc, True)
+        MaterialComponent(Nuclide(isotope, suppress_warning=True), conc, True)
 
 
 def test_mat_comp_init_warn():
     with pytest.warns(DeprecationWarning):
-        MaterialComponent(Isotope("1001.80c", suppress_warning=True), 0.1)
+        MaterialComponent(Nuclide("1001.80c", suppress_warning=True), 0.1)
 
 
 def test_material_comp_fraction_setter():
-    comp = MaterialComponent(Isotope("1001.80c", suppress_warning=True), 0.1, True)
+    comp = MaterialComponent(Nuclide("1001.80c", suppress_warning=True), 0.1, True)
     comp.fraction = 5.0
     assert comp.fraction == pytest.approx(5.0)
     with pytest.raises(ValueError):
@@ -109,7 +109,7 @@ def test_material_comp_fraction_setter():
 
 
 def test_material_comp_fraction_str():
-    comp = MaterialComponent(Isotope("1001.80c", suppress_warning=True), 0.1, True)
+    comp = MaterialComponent(Nuclide("1001.80c", suppress_warning=True), 0.1, True)
     str(comp)
     repr(comp)
 
@@ -123,7 +123,7 @@ def test_material_update_format():
     print(material.format_for_mcnp_input((6, 2, 0)))
     assert "8016" in material.format_for_mcnp_input((6, 2, 0))[0]
     # addition
-    isotope = Isotope("2004.80c", suppress_warning=True)
+    isotope = Nuclide("2004.80c", suppress_warning=True)
     with pytest.deprecated_call():
         material.material_components[isotope] = MaterialComponent(isotope, 0.1, True)
         print(material.format_for_mcnp_input((6, 2, 0)))
@@ -208,29 +208,29 @@ def test_bad_init(line):
 
 # test isotope
 def test_isotope_init():
-    isotope = Isotope("1001.80c")
+    isotope = Nuclide("1001.80c")
     assert isotope.ZAID == "1001"
     assert isotope.Z == 1
     assert isotope.A == 1
     assert isotope.element.Z == 1
     assert isotope.library == "80c"
     with pytest.raises(ValueError):
-        Isotope("1001.80c.5")
+        Nuclide("1001.80c.5")
     with pytest.raises(ValueError):
-        Isotope("hi.80c")
+        Nuclide("hi.80c")
 
 
 def test_isotope_metastable_init():
-    isotope = Isotope("13426.02c")
+    isotope = Nuclide("13426.02c")
     assert isotope.ZAID == "13426"
     assert isotope.Z == 13
     assert isotope.A == 26
     assert isotope.is_metastable
     assert isotope.meta_state == 1
-    isotope = Isotope("92635.02c")
+    isotope = Nuclide("92635.02c")
     assert isotope.A == 235
[…]
