### 2024-09-12 8ff4bbc
Started prototype for filtering isotopes.
