### 2024-09-12 1b385a5
Finished concept of finding things.
