### 2024-09-12 068d8b0
Deleted dead code.
