### 2024-09-12 34a9ed7
Simplified fancy name to be string only.
