### 2024-09-12 2630231
Removed all material_components.
