### 2024-09-12 17825c5
cleaned up small bugs.
