### 2024-09-12 3fe4847
Simplified tests for new interface.
