### 2024-09-12 d3e6007
Played around with hypothesis and PBT.
