### 2024-09-13 521fcf9
Ignore non-sense isotopes, and helped debugging.
