### 2024-09-13 6e7b21b
Fixed bad format re for nuclide.
