### 2024-09-13 c639e23
Handled case of bad string.
