### 2024-09-13 7d78397
Changed how A is sampled to be more physically accurate.
