### 2024-09-15 4629e78
Made isotope  fancy name more exhaustive.
