### 2024-09-15 c5337bf
Fixed issue of dissapearing isomer.
