### 2024-09-15 c1b567d
Updated test matrix and moved to 3.12 as preferred.
