+++ .github/workflows/main.yml
@@ -10,7 +10,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
 
     steps:
       - uses: actions/checkout@v4
@@ -36,15 +36,13 @@ jobs:
       - run: pip install --user dist/*.tar.gz
       - run: pip install --user . montepy[test]
       - run: pip install --user . montepy[doc]
-        if: ${{ matrix.python-version != '3.8'}}
       - run: pip install --user . montepy[format]
       - run: pip install --user . montepy[build]
       - run: pip install --user . montepy[develop]
-        if: ${{ matrix.python-version != '3.8'}}
       - run: pip freeze
       - name: Upload build artifacts
         uses: actions/upload-artifact@v4.3.1
-        if: ${{ matrix.python-version == '3.9' }}
+        if: ${{ matrix.python-version == '3.12' }}
         with: 
            name: build
            path: dist/*
@@ -70,19 +68,19 @@ jobs:
       - run: coverage xml
         if:  ${{ success() || failure() }}
       - name: Upload test report
-        if: ${{ matrix.python-version == '3.9' && (success() || failure() )}}
+        if: ${{ matrix.python-version == '3.12' && (success() || failure() )}}
         uses: actions/upload-artifact@v4
         with:
           name: test
           path: test_report.xml
       - name: Upload coverage report
-        if: ${{ matrix.python-version == '3.9' && (success() || failure() )}}
+        if: ${{ matrix.python-version == '3.12' && (success() || failure() )}}
         uses: actions/upload-artifact@v4
         with:
           name: coverage
           path: coverage.xml
       - name: Coveralls GitHub Action
-        if: ${{ matrix.python-version == '3.9' && (success() || failure() )}}
+        if: ${{ matrix.python-version == '3.12' && (success() || failure() )}}
         uses: coverallsapp/github-action@v2
         with:
           file: coverage.xml
