### 2024-09-15 395befe
Started testing against numpy 1 and 2.
