+++ .github/workflows/main.yml
@@ -52,7 +52,8 @@ jobs:
     permissions: write-all
     strategy:
       matrix:
-        python-version: ["3.8", "3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.9", "3.10", "3.11", "3.12", "3.13"]
+        numpy-version: ["1.0", "2.0"]
     
     steps:
       - uses: actions/checkout@v4
@@ -60,8 +61,11 @@ jobs:
         uses: actions/setup-python@v5
         with: 
           python-version: ${{ matrix.python-version }}
+      - run: pip install numpy~=${{ matrix.numpy-version }}
+        name: install specific numpy version
       - run: pip install --user . montepy[test]
       - run: pip install --user . montepy[build]
+      - run: pip freeze
       - run: coverage run -m pytest --junitxml=test_report.xml
       - run: coverage report
         if:  ${{ success() || failure() }}
