### 2024-09-16 307ae3e
Limited A multiplier to being more physical.
