### 2024-09-16 e34d83c
Removed material_components from tests.
