### 2024-09-16 92facc4
Made a lot of data_inputs as top level objects.
