+++ montepy/__init__.py
@@ -10,17 +10,29 @@ You will receive an MCNP_Problem object that you will interact with.
 from . import input_parser
 from . import constants
 import importlib.metadata
-from .input_parser.input_reader import read_input
-from montepy.cell import Cell
-from montepy.mcnp_problem import MCNP_Problem
+
+# data input promotion
+
 from montepy.data_inputs.material import Material
 from montepy.data_inputs.transform import Transform
+from montepy.data_inputs.nuclide import Nuclide
+from montepy.data_inputs.element import Element
+
+# geometry
 from montepy.geometry_operators import Operator
 from montepy import geometry_operators
-from montepy.input_parser.mcnp_input import Jump
-from montepy.particle import Particle
 from montepy.surfaces.surface_type import SurfaceType
+
+# input parser
+from montepy.input_parser.mcnp_input import Jump
+from .input_parser.input_reader import read_input
+
+# top level
+from montepy.particle import Particle
 from montepy.universe import Universe
+from montepy.cell import Cell
+from montepy.mcnp_problem import MCNP_Problem
+
 import montepy.errors
 import sys
 
