### 2024-09-16 e6b2bcb
Made elements print by reverse order and truncated.
