### 2024-09-16 b45e535
Made iterator of material to hide valuenode.
