### 2024-09-16 b61b2d3
Added elements set to cache info.
