### 2024-09-16 5e0ff5e
Made easier way to add new nuclides.
