### 2024-09-16 d12ec1a
Made it possible to find nuclide and element in material.
