### 2024-09-16 718bfa2
Made Element truly immutable with slots.
