+++ montepy/data_inputs/nuclide.py
@@ -42,7 +42,17 @@ _ZAID_A_ADDER = 1000
 
 class Nucleus:
 
-    __slots__ = "_element", "_a", "_meta_state"
+    __slots__ = "_element", "_A", "_meta_state"
+
+    #                   Cl-52      Br-101     Xe-150      Os-203    Cm-251     Og-296
+    _BOUNDING_CURVE = [(17, 52), (35, 101), (54, 150), (76, 203), (96, 251), (118, 296)]
+    _STUPID_MAP = {
+        "95642": {"_is_metastable": False, "_meta_state": None},
+        "95242": {"_is_metastable": True, "_meta_state": 1},
+    }
+    """
+    Points on bounding curve for determining if "valid" isotope
+    """
 
     def __init__(
         self,
@@ -56,10 +66,10 @@ class Nucleus:
             parts = ZAID.split(".")
             try:
                 assert len(parts) <= 2
-                int(parts[0])
+                ZAID = int(parts[0])
             except (AssertionError, ValueError) as e:
                 raise ValueError(f"ZAID: {ZAID} could not be parsed as a valid isotope")
-            new_vals = self._parse_zaid(int(self._ZAID))
+            new_vals = self._parse_zaid(int(ZAID))
             for key, value in new_vals.items():
                 setattr(self, key, value)
         elif element is not None:
@@ -73,7 +83,7 @@ class Nucleus:
             if not isinstance(Z, int):
                 raise TypeError(f"Z number must be an int. {Z} given.")
             self._element = Element(Z)
-        self._handle_stupid_legacy_stupidity()
+        self._handle_stupid_legacy_stupidity(ZAID)
         if ZAID:
             return
         if A is not None:
@@ -88,10 +98,139 @@ class Nucleus:
             self._meta_state = meta_state
         else:
             self._meta_state = 0
-        if not isinstance(library, str):
-            raise TypeError(f"Library can only be str. {library} given.")
-        self._library = Library(library)
-        self._ZAID = str(self.get_full_zaid())
+
+    def _handle_stupid_legacy_stupidity(self, ZAID):
+        # TODO work on this for mat_redesign
+        if ZAID in self._STUPID_MAP:
+            stupid_overwrite = self._STUPID_MAP[self.ZAID]
+            for key, value in stupid_overwrite.items():
+                setattr(self, key, value)
+
+    @property
+    def ZAID(self):
+        """
+        The ZZZAAA identifier following MCNP convention
+
+        :rtype: int
+        """
+        meta_adder = 300 + 100 * self.meta_state if self.is_metastable else 0
+        return self.Z * _ZAID_A_ADDER + self.A + meta_adder
+
+    @property
+    def Z(self):
+        """
+        The Z number for this isotope.
+
+        :returns: the atomic number.
+        :rtype: int
+        """
+        return self._element.Z
+
+    @make_prop_pointer("_A")
+    def A(self):
+        """
+        The A number for this isotope.
+
+        :returns: the isotope's mass.
+        :rtype: int
+        """
+        pass
+
+    @make_prop_pointer("_element")
+    def element(self):
+        """
+        The base element for this isotope.
+
+        :returns: The element for this isotope.
+        :rtype: Element
+        """
+        pass
+
+    @property
+    def is_metastable(self):
+        """
+        Whether or not this is a metastable isomer.
+
+        :returns: boolean of if this is metastable.
+        :rtype: bool
+        """
+        return bool(self._meta_state)
+
+    @make_prop_pointer("_meta_state")
+    def meta_state(self):
+        """
+        If this is a metastable isomer, which state is it?
+
+        Can return values in the range [1,4] (or None). The exact state
+        number is decided by who made the ACE file for this, and not quantum mechanics.
+        Convention states that the isomers should be numbered from lowest to highest energy.
+
[…]
