### 2024-09-16 543c764
Updated equality to new iterator.
