### 2024-09-16 7330ce9
Debugged nuclide str.
