### 2024-09-17 1a97f28
Fixed ZAID to be treated as an int actually.
