### 2024-09-17 d993776
Test only that material_component is deprecated.
