### 2024-09-17 5fe7a67
Fixed update_values to use new data_structure.
