### 2024-09-17 228acbf
Brought back new thermal test for new update_pointer.
