+++ tests/test_material.py
@@ -525,9 +525,10 @@ def test_thermal_scattering_format_mcnp():
     in_str = "M20 1001.80c 0.5 8016.80c 0.5"
     input_card = Input([in_str], BlockType.DATA)
     material = Material(input_card)
-    material.update_pointers([card])
+    material.thermal_scattering = card
+    card._parent_material = material
     material.thermal_scattering.thermal_scattering_laws = ["grph.20t"]
-    assert card.format_for_mcnp_input((6, 2, 0)) == ["Mt20 grph.20t "]
+    card.format_for_mcnp_input((6, 2, 0)) == ["Mt20 grph.20t "]
 
 
 def test_thermal_str():
