### 2024-09-17 edc3181
Exposed nucleus as a property.
