### 2024-09-17 f33283d
Moved nuclide equality to mostly nucleus.
