### 2024-09-17 9252022
Implemented searching for materials by components.
