### 2024-09-17 e30b570
Added enum for material libraries types.
