+++ montepy/surfaces/surface_type.py
@@ -1,9 +1,9 @@
 # Copyright 2024, Battelle Energy Alliance, LLC All Rights Reserved.
-from enum import unique, Enum
+from enum import unique, StrEnum
 
 
-# @unique
-class SurfaceType(str, Enum):
+@unique
+class SurfaceType(StrEnum):
     """
     An enumeration of the surface types allowed.
 
