### 2024-09-17 6e0b614
Added default_libraries to material class.
