+++ montepy/mcnp_problem.py
@@ -317,7 +317,7 @@ class MCNP_Problem:
                         if isinstance(obj, Material):
                             self._materials.append(obj, False)
                         if isinstance(obj, transform.Transform):
-                            self._transforms.append(obj)
+                            self._transforms.append(obj, False)
                     if trailing_comment is not None and last_obj is not None:
                         obj._grab_beginning_comment(trailing_comment)
                         last_obj._delete_trailing_comment()
