### 2024-09-18 c0946a7
Ensured that all problem collections are linked.
