### 2024-09-18 aa94477
Improved deepcopy to not keep pointing to old thing.
