+++ pyproject.toml
@@ -18,7 +18,6 @@ keywords = ["MCNP", "neutronics", "imcnp", "input file", "monte carlo", "radiati
 license = {file="LICENSE"}
 classifiers = [
        "Development Status :: 4 - Beta",
-       "Programming Language :: Python :: 3 :: Only",
        "Intended Audience :: Science/Research",
        "Intended Audience :: Developers",
        "License :: OSI Approved :: MIT License",
@@ -27,7 +26,12 @@ classifiers = [
        "Topic :: Scientific/Engineering :: Physics",
        "Topic :: Scientific/Engineering",
        "Topic :: Scientific/Engineering :: Human Machine Interfaces",
-
+       "Programming Language :: Python :: 3 :: Only",
+       "Programming Language :: Python :: 3.9",
+       "Programming Language :: Python :: 3.10",
+       "Programming Language :: Python :: 3.11",
+       "Programming Language :: Python :: 3.12",
+       "Programming Language :: Python :: 3.13"
 ]
 dependencies = [
 	"numpy>=1.18",
