### 2024-09-23 28186e7
Actually test reading from a file stream.
