+++ tests/test_integration.py
@@ -1109,7 +1109,7 @@ def test_read_write_cycle(file):
     problem.write_problem(fh)
     fh.seek(0)
     # test valid syntax
-    new_problem = montepy.MCNP_Problem("foo")
+    new_problem = montepy.read_input(fh)
     # verify lines are similar
     fh.seek(0)
     lines = [line.rstrip() for line in fh]
