### 2024-09-23 e470d30
Avoided opening an already open stream.
