### 2024-09-23 05f2a86
Make stringIO unclosable.
