+++ tests/test_integration.py
@@ -1106,6 +1106,8 @@ def test_read_write_cycle(file):
     problem = montepy.read_input(file)
     SKIPPERS = _SKIP_LINES.get(str(file), {})
     fh = io.StringIO()
+    # make string unclosable to keep open after reading.
+    fh.close = lambda: None
     problem.write_problem(fh)
     fh.seek(0)
     # test valid syntax
