### 2024-09-23 b1fb135
Made meta class to make immutables singletons.
