### 2024-09-23 8c84a10
Made elements and nucleus singletons.
