### 2024-09-23 a85ff52
Moved get_from_fancy_name into init
