### 2024-09-23 af541d6
Optimized equality for singleton objects
