### 2024-09-23 37e5ce3
Implemented prototype of set logic for numbered_object_collection.
