### 2024-09-23 #205 @kellyrowland
@MicahGale @munkm @jpmorgan98 @cmarmo just FYI I will be out of office starting September 30 and will return on October 21.
