### 2024-09-24 f4be3f9
Made multi-format multi-resolution favicons.
