### 2024-09-24 2331d1f
Moved GHA deploy to using python check version.
