+++ .github/workflows/deploy-alpha.yml
@@ -39,8 +39,8 @@ jobs:
       - name: Get Version
         id: get_version
         run: echo "version=`python -m setuptools_scm`" >> $GITHUB_OUTPUT
-      - name: Verify that this is a non-dev release
-        run: .github/scripts/check_version.sh ${{ steps.get_version.outputs.version }}
+      - name: Verify that this is a non-dev alpha release
+        run: .github/scripts/check_version.py --alpha 
       - run: python -m build .
       - name: Sign the dists with Sigstore
         uses: sigstore/gh-action-sigstore-python@v2.1.1
