### 2024-09-24 30e7d18
Made NumberedObject more like a set.
