### 2024-09-25 9dd385b
Initial implementation for clone.
