### 2024-09-25 5ad0b4c
Made mcnp problem repr actually meaningful and exhaustive.
