### 2024-09-25 26ce03d
meta-programmed loading all collections.
