### 2024-09-25 ea9bc60
Fixed bugs with problem deepcopy.
