### 2024-09-25 60ea197
Updated problem ref key in getstate.
_ also avoided bug with materials not being made after data_inputs
