### 2024-09-25 b044eb6
Added #469 to changelog.
