### 2024-10-09 idaholab/MontePy#433 @tjlaboss
https://github.com/alexhagen/wig (http://alexhagen.github.io/wig/)
