### 2024-10-09 idaholab/MontePy#433 @MicahGale
https://ieeexplore.ieee.org/document/10655039
