### 2024-10-16 idaholab/MontePy#572 @jpmorgan98
Unit test failure (`test_version`) (v0.4.1)
**Describe the bug**

Failed `TestVersion.test_version` when running comprehensive test suite.

**To Reproduce**

`cd MontePy`
`pytest`

**Error Message (if any)**
```
================================================== FAILURES ==================================================
__________________________________________ TestVersion.test_version __________________________________________

self = <tests.test_version.TestVersion testMethod=test_version>

    def test_version(self):
        print(os.path.join("montepy", "_version.py"))
>       self.assertTrue(os.path.exists(os.path.join("montepy", "_version.py")))
E       AssertionError: False is not true

tests/test_version.py:10: AssertionError
-------------------------------------------- Captured stdout call --------------------------------------------
montepy/_version.py
========================================== short test summary info ===========================================
FAILED tests/test_version.py::TestVersion::test_version - AssertionError: False is not true
```

**Version**

 - Version 0.4.1 with optional testing dependencies (`pip install 'montepy[test]'`)
 - Python 3.12.3

**Additional context**

[Part of pyOpenSci review!](https://github.com/pyOpenSci/software-submission/issues/205)
