### 2024-10-16 #205 @jpmorgan98
4
---

A very well developed package with well written clear documentation. It always makes me happy to see open source work supporting radiation transport! I am not the most avid MCNP user but have had my fair share of evenings lost to unknown syntax errors in my input decks, so it's exciting to see a project address that in a well developed and thought out way.

Doc strings are very well written and integrate nicely into a well thought out documentation cite. They seem to be associated with every function at all depths.

I have some comments expanding on unchecked boxes above.

##### Itemized issues
- [ ] Version test not passing on local machine (github issue [572](https://github.com/idaholab/MontePy/issues/572))
- [ ] Demo notebook failing on local machine (github issue [571](https://github.com/idaholab/MontePy/issues/571))
- [ ] Specific requested badges (specifically: Docs building, repostatus.org badge, Python versions supported)
- [ ] Vignette from README not working locally. I have no `foo.imcnp` to compare too, maybe putting one in the demo directory and linking it might be nice
- [ ] Citation information needed at the bottom of the README. I am guessing this will be the JOSS paper once approved so just an acknowledgment that you'll do this once MontePy is accepted should suffice to check this off.
- [ ] Noting similar packages. You mention in the paper ARMI, WATTS, and MCNPy but they should also be mentioned on your README with a quick note about how MontePy is different.

**NOTE** I do not have access to an installation of MCNP. So for the purposes of this review I cannot validate that the output decks that MontePy provides, work.

You mention in the paper some of the other packages that do this but are less general purpose. I think [PyNE](http://pyne.io/pyapi/mcnp.html?highlight=mcnp#module-pyne.mcnp) also has some limited functionality to do this, tho I could be mistaken. Also another thing of note is that the other LANL Monte Carlo rad transport app [MCATK has a built in Python interface](https://www.kns.org/files/int_paper/paper/MC2017_2017_2/P176S02-10TrahanT.pdf) which only further demonstrates the need for something general purpose like this for MCNP.

I also noticed that you just did a version release to 0.5.0, first off congrats, second should we change which version we are reviewing?

Great job again on such a well developed package!
