### 2024-10-16 idaholab/MontePy#572 @MicahGale
Currently the version string is generated by `setuptools_scm` on deploy by writing `montepy/_version.py`. This inherently means that `_version.py` can't be tracked and isn't cloned. Due to this [`montepy/__init__.py`](https://github.com/idaholab/MontePy/blob/develop/montepy/__init__.py) has three ways to determine the version. This test is trying to verify all three methods work. It assumes that `_version.py` is there. The weird thing is that in CI I can't see where this file is coming from, and it shouldn't be passing?

This test just needs to be updated to ensure that `setuptools_scm` generates that file.

A temporary workaround would be to run `python -m build .`
