### 2024-10-16 idaholab/MontePy#433 @MicahGale
We should also discuss PyNE. I'm thinking we discuss the more developed ones, and then just make a bullet list of the less complete ones and just say they don't have the same features.
