### 2024-10-16 #205 @MicahGale
Thank you @jpmorgan98 for your review and feedback!

re:version

I have been updating the submitted version with every release since this was started. However, I'm not sure if I should keep doing this though. Depending on what branch you were looking at you were either review 0.4.1 or 0.4.2devNNN (dev versions default to a patch release, so 0.4.2.devNNN became 0.5.0 on release because it was a minor release.) So I'll leave it to @kellyrowland on how we should deal with this once she is able to return.

re: Competitors.

Yes PyNE is definitely worth mentioning in the paper, and not just because Paul Wilson is my advisor. Since writing the first draft of this we have come across a lot more packages, and opened a lazy issue ([433](https://github.com/idaholab/MontePy/issues/433)) to track these. This should be finalized prior to finalizing this review, right?

re: Validating if the MCNP input files work

Can't you just run `python -m montepy -c ` :P. Yes this is a bit of a challenge sometimes and would require tedious manual review of the input files against the manual to see if the file makes sense. We do actually have a [cyclical test](https://github.com/idaholab/MontePy/blob/develop/tests/test_integration.py#L1104) where a file read-written-read to make sure nothing  breaks. Obviously this isn't perfect, that it does catch a lot of low hanging fruit.
