### 2024-10-16 idaholab/MontePy#433 @MicahGale
I started something like this in #573. I think slightly more detail than that is probably appropriate?
