### 2024-10-23 #205 @kellyrowland
Fair question on versioning - my experience with JOSS is that folks usually submit version X.Y which is reviewed, and then when any edits have gone in, we have the submitter cut a new release X.Z which contains the edits and then that gets an associated archive that goes with the publication. So, the JOSS publication is associated with some specific version of the software.

@MicahGale is there a pinned branch or version that you would preferably have the reviewers look at for being tagged along with the manuscript? For the JOSS review process it's generally not the case that the codebase gets a major overhaul for functionality, interface, etc. (often just manuscript edits), so any recent tag or version would be sufficient IMO. Let me know if you have any questions.
