### 2024-10-24 #205 @MicahGale
I think it should be fine if everyone reviews `develop`. The latest release was a minor one that added a new feature, and fixed a few bugs. We won't do a major release during this review that would remove any current functionality. So I think just keeping up with the current releases should be fine, and will reduce the amount of duplicate issues that reviewers find.
