### 2024-11-04 #205 @MicahGale
I wonder if @munkm is at the ANS winter conference this week? The office is a bit of a ghost town for us right now due to it.
