### 2024-11-18 #205 @kellyrowland
Thanks for the note! I'm more active in the HPC space than the nuclear engineering space these days, so not so plugged into the immediate ANS calendar. 🙂
