### 2025-01-27 #205 @munkm
---

This is a really nice package, and I like the design approach that you've used. I wish I had been able to use this when I was first building MCNP input (decks). 🙂 I think your narrative documentation reads nicely and you do a great job demonstrating the functionality of the package. This tool is useful, fits in the ecosystem, and has the potential to support anybody's research that uses MCNP!

A few notes related to unchecked boxes in my review:
- It looks to me in your badges that you do include a test coverage and a docs deploy badge, but not a test status badge (passing/not passing). I do see your test results from the most recent merged PR.
- You may want to include install instructions for optional dependencies for interested users/developers or some language in your documentation that to run some examples you will need extra dependencies. For example, you have a .ipynb dile in /demos/ but you do not list jupyter notebooks as an optional dependency. I see that in your pyproject.toml file you have a demos configuration, but this isn't documented in your installation instructions that I saw.
- To build off of @jpmorgan98 's comment about PyNE features, I think adding a reference to PyNE to the paper would be good too.
- In the paper you mention MCNPtools, but do not cite it. The MCNPtools repo does have a preferred citation: https://github.com/lanl/mcnptools
- In reviewing the examples and documentation I didn't see documentation on how to set boundary conditions (I did see the `is_reflecting` property and `is_white_boundary` for surfaces, as well as `periodic_surface`. You do a great job of describing how to bring surfaces together to form cells, but an example on setting these would be helpful.
- In reviewing the examples I had a question: in the example above `Setting and Modifying Geometry` in the User Guide you specify building a cylindrical geometry using `AxisPlane` to form two truncating planes. You don't use any default orientation for these in the example. Is it assumed to be x on writing (is there a default?) ? Or is there a requirement to specify a plane before writing? I think a note to document this and the preferred method to set it would be helpful.

@kellyrowland the authors do extensively document the API for their functionality, and the narrative docuumentation describes how cells, surfaces, and material properties are set. However, not every function--for example surface modules like `cylinder_par_axis` and `surface_builder`--has an example in the documentation. I am not sure how much value it would add to request this, but it is part of the checklist for this review. It is my opinion that the documentation covers the general functionality. Can you please advise on what the best course of action is here?
