### 2025-01-27 #205 @MicahGale
Thank you for this review @munkm! I have opened: https://github.com/idaholab/MontePy/issues/646 to track resolution of your comments.

A few questions / comments I had:

> You may want to include install instructions for optional dependencies for interested users/developers or some language in your documentation that to run some examples you will need extra dependencies. For example, you have a .ipynb dile in /demos/ but you do not list jupyter notebooks as an optional dependency. I see that in your pyproject.toml file you have a demos configuration, but this isn't documented in your installation instructions that I saw.

Do you think a new subsection of this section would be a good place for these installation options? https://www.montepy.org/en/stable/starting.html#installing

> In reviewing the examples I had a question: in the example above Setting and Modifying Geometry in the User Guide you specify building a cylindrical geometry using AxisPlane to form two truncating planes. You don't use any default orientation for these in the example. Is it assumed to be x on writing (is there a default?) ? Or is there a requirement to specify a plane before writing? I think a note to document this and the preferred method to set it would be helpful.

Good question. I think this is actually an invalid example. IIRC there is no default orientation, or really type (e.g., `PX`), for `AxisPlane` so I think on export this example would lead to an `IllegalState` error. I will fix that demo.

> To build off of @jpmorgan98 's comment about PyNE features, I think adding a reference to PyNE to the paper would be good too. In the paper you mention MCNPtools, but do not cite it. The MCNPtools repo does have a preferred citation: https://github.com/lanl/mcnptools

Thanks for catching that we missed those citations. We also have an issue where we have tracked all competitors we missed in the first draft: https://github.com/idaholab/MontePy/issues/433. The plan is to write a section similar to the "competitors" section in the README. We'll make sure to get to that issue before we finish this review process.

Finally:

> This tool is useful, fits in the ecosystem, and has the potential to support anybody's research that uses MCNP!

This is great, and what we were really going for!
