### 2025-01-27 #205 @munkm
Thanks @MicahGale !!! Feel free to tag me in the issue or associated PRs if you'd like.

> Do you think a new subsection of this section would be a good place for these installation options? https://www.montepy.org/en/stable/starting.html#installing

Yes I think this would be a great place for it! Or if you prefer to keep that section light you could put it in the development guide install instructions and link to it from the getting started guide.

>  I will fix that demo.

Ok, great! I wasn't quite sure how to set them based on that, so I think it will be a nice improvement!
