### 2025-01-27 #205 @munkm
I had one last thought @MicahGale -- I know the scope of this project is to support the full MCNP manual eventually. If you do plan to support tallies and eigenvalue configurations (I may have missed it if you said it either way?), you may want to include something in your API documentation and in the README that there is a path for these in the future.
