### 2025-01-28 #205 @MicahGale
Good point.

Yes those are planned for one day https://github.com/idaholab/MontePy/issues/11, https://github.com/idaholab/MontePy/issues/21, https://github.com/idaholab/MontePy/issues/56 (once I have enough time).
