### 2025-01-29 #205 @MicahGale
> In the paper you mention MCNPtools, but do not cite it. The MCNPtools repo does have a preferred citation: https://github.com/lanl/mcnptools

I doubled checked on this, we do cite Bates, et al. The citation is one sentence after we mention MCNPtools, would elsewhere be clearer for having the citation? @munkm
