### 2025-01-29 #205 @MicahGale
@kellyrowland I just wanted to double check you saw @munkm's question [above](https://github.com/pyOpenSci/software-submission/issues/205#issuecomment-2616561493):

> @kellyrowland the authors do extensively document the API for their functionality, and the narrative docuumentation describes how cells, surfaces, and material properties are set. However, not every function--for example surface modules like cylinder_par_axis and surface_builder--has an example in the documentation. I am not sure how much value it would add to request this, but it is part of the checklist for this review. It is my opinion that the documentation covers the general functionality. Can you please advise on what the best course of action is here?
