### 2025-01-31 #205 @munkm
> > In the paper you mention MCNPtools, but do not cite it. The MCNPtools repo does have a preferred citation: https://github.com/lanl/mcnptools
>
> I doubled checked on this, we do cite Bates, et al. The citation is one sentence after we mention MCNPtools, would elsewhere be clearer for having the citation? [@munkm](https://github.com/munkm)

I think where you have it is fine. I think I was focusing on first sentence in the review, wrote the feedback, and when I restarted didn't realize the next sentence had the reference. I'm happy for it to remain unchanged.
