### 2025-03-18 #205 @MicahGale
@kellyrowland it sounds like we've got all the review comments addressed.

What are the next steps? I think we might be ready for [step 6: package acceptance](https://www.pyopensci.org/software-peer-review/our-process/review-timeline.html#package-acceptance).
