### 2025-03-19 #205 @kellyrowland
yep, I think it's about that time. 🚀

@MicahGale please create a new Github release with the changes if you haven't already; ideally since we're going to do the companion JOSS paper this will be the release and version to which the archive and DOI that you'll create will match up with so that's all done in one fell swoop. I'll note here that we ask authors to put the archive on somewhere like Zenodo and then post the DOI (that can go in the JOSS issue and we can follow up there on JOSS specifics).

@cmarmo are there any steps for you to take now?
