### 2025-03-19 #205 @MicahGale
@kellyrowland I found this [checklist/template](https://www.pyopensci.org/software-peer-review/how-to/editors-guide.html#how-to-accept-a-package-into-the-pyopensci-ecosystem) in the editor guide for you. It looks like there's just a few things to wrap up.

I am checking with our organization admin on github for adding the Zenodo integration, and will hopefully get that figured out in the next few days. I'll be doing the release to finalize this review later today (that will be version 0.5.5).

Update: 0.5.5 has been released to PyPI. TODO list:

* [x] enable Zenodo
* [x] publish draft release on GH
* [x] Publish to conda-forge
