### 2025-03-20 #205 @MicahGale
MontePy has been published on Zenodo: [![DOI](https://zenodo.org/badge/DOI/10.11578/dc.20240115.1.svg)](https://doi.org/10.11578/dc.20240115.1)

Note: however, we had a previous DOI for MontePy, so the DOI link does not go to Zenodo. The [Zenodo archive is here](https://zenodo.org/records/15058701). Is this an issue?
