### 2025-03-20 #205 @cmarmo
> @cmarmo are there any steps for you to take now?

Not really... :) a part thank you all for all your work!
I am not sure though that the zenodo archive is necessary as MontePy was already declared on the  U.S. Department of Energy Office of Scientific and Technical Information archive...
I guess it's up to the author to decide which one they prefer keeping updated.
