### 2025-04-30 #205 @lwasser
hi everyone 👋🏻 i wasn't able to find the acceptance checklist here to finish updating the YAML file and close this review. i'm going to paste it here. (UNLESS I MISSED IT) and then ask for some help filling it out 🚀 (We parse these issues so this information is needed!!)

---

🎉 MontePy has been approved by pyOpenSci! Thank you @MicahGale, @tjlaboss for submitting MontePy and many thanks to @munkm  @jpmorgan98 for reviewing this package! 😸

- [x] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

<IF JOSS SUBMISSION>

- [x] <Editor Task> Once the JOSS issue is opened for the package, we strongly suggest that you subscribe to issue updates. This will allow you to continue to update the issue labels on this review as it goes through the JOSS process.

🎉 Congratulations! You are now published with both JOSS and pyOpenSci! 🎉
<IF JOSS SUBMISSION/>

---
