### 2025-04-30 #205 @lwasser
@MicahGale @kellyrowland can you kindly fill me in here.
✅ I think the package was accepted on 19 march. But I also see that it went through JOSS.
✅  it looks like zenodo was added to the top and the version is 0.5.5 (thank you Micah!
✅  it also looks like you filled out the post survey -- THANK YOU. ✨

I think we are in good shape. i've added the JOSS doi, version accepted etc to the yaml header!
