### 2025-04-30 #205 @lwasser
@MicahGale and EVERYONE involved in this review, if you are interested in joining our slack (if you aren't already there!) please email me at leah at pyopensci.org and i'll shoot you a link! Thank you for all of your collective time in running this review!! 🚀 i hope that you are all well and taking good care of yourselves during these tricky times. 💓
