### 2025-05-01 #205 @lwasser
hi @MicahGale of course!! Blog posts are definitely still a thing and we welcome you to submit one!

* [Here is an overview of how to submit one to us](https://www.pyopensci.org/handbook/organization/social-media-blog.html#blog-post-submission-process) .
* [Here is a blog post submitted a few months ago as an example. ](https://www.pyopensci.org/blog/quadratik.html)

If you have any questions, feel free to ask here OR you can also ask in Slack in our #software-review channel.
I'm open to updating our docs if you have questions so please do ask if things in the docs could be more clear!
